Function to apply a single vaccination activity to the population

If the vaccination activity has non-missing entries for both coverage and doses, the coverage is used - there is no test whether the coverage and doses information are consistent with the target population size in the vip_population object to which the vaccination activity is applied.

apply_vacc(
  pop_df,
  region,
  year,
  age_first = 0,
  age_last = Inf,
  coverage = double(),
  doses = NA_real_,
  targeting = "random"
)

Arguments

pop_df

population dataframe object such as created by function 'vip_population'

region

region of the vaccination activity

year

year of the vaccination activity

age_first

age of the youngest age group targeted

age_last

age of the oldest age group targeted

coverage

proportion of the population to be vaccinated in the activity

doses

number of doses available to vaccinate the target population

targeting

character to determine how successive activities are targeted. Valid values are "random", "correlated", "targeted". For targeting = "random" (the default), allocation of vaccine is random within the population, therefore the resulting coverage will be smaller than the sum of the coverages. For targeting = "correlated", there is a 100% correlation between who will get the vaccine in either vaccination activitiy, the resulting coverage is simply the larger of the two inputs. For targeting = "targeted", doses are targeted at unvaccinated people, resulting in the sum of both coverages, though capped by 1 (full coverage).

Value

pop_df: the supplied population dataframe object with updated immunity to reflect the vaccination activity.

Author

Tini Garske