Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. These data were originally described in Friedman (1991) and Breiman (1996). For details, see sklearn.datasets.make_friedman1.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

References

Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The Annals of Statistics 19 (1), pages 1-67.

Examples

gen_friedman()
#>             y           x1         x2          x3         x4           x5
#> 1   23.656921 0.8585893337 0.83277146 0.976175347 0.88049776 0.4933718622
#> 2   18.643193 0.5093839013 0.46917252 0.671394317 0.67293058 0.8863800692
#> 3   14.441017 0.6913609528 0.44104079 0.464803080 0.32966965 0.5662391454
#> 4   15.840893 0.4404512052 0.53014473 0.328135147 0.55371357 0.5962013067
#> 5   13.223209 0.2745307968 0.32955744 0.886772868 0.51460093 0.4405027200
#> 6    7.441523 0.1394910647 0.83690150 0.856671865 0.10447918 0.0731950905
#> 7   14.021413 0.4747850548 0.05433750 0.028965648 0.68141135 0.4111418726
#> 8   16.977932 0.6787334885 0.95498762 0.109910289 0.15130576 0.7270049329
#> 9   10.747124 0.4866766350 0.16536440 0.656012750 0.35471463 0.8435586512
#> 10  17.591894 0.1083969162 0.85478324 0.951910495 0.95992138 0.2199263778
#> 11   7.544029 0.6706628674 0.13130213 0.357360833 0.24143085 0.4397629711
#> 12  10.204395 0.4743734493 0.46372246 0.472532317 0.32830772 0.1060734710
#> 13  17.617884 0.8636652103 0.66398833 0.814777710 0.56747620 0.0505345890
#> 14   8.376512 0.4951284917 0.36504793 0.462451591 0.12384373 0.3366554687
#> 15  20.781944 0.8932647803 0.67154270 0.837737901 0.49869163 0.8571355718
#> 16  18.525845 0.2484758606 0.13309245 0.006623439 0.85631812 0.7888293739
#> 17  10.362254 0.1880515858 0.72135466 0.295536251 0.30228006 0.4960319682
#> 18  17.921912 0.6389388076 0.58116697 0.575176170 0.69947411 0.3228566304
#> 19  14.041015 0.4431636555 0.13014876 0.950826771 0.63629729 0.3545236189
#> 20  19.335793 0.5978760927 0.85557827 0.599125821 0.81173009 0.2008812900
#> 21   6.508853 0.5402609040 0.11568335 0.822221182 0.06891674 0.3956458082
#> 22  21.923718 0.6818523891 0.49731483 0.760869240 0.99248806 0.4004982647
#> 23   7.206972 0.0144383404 0.57732297 0.076940237 0.23390536 0.2143567074
#> 24  22.773911 0.8716980382 0.63333918 0.037994279 0.37095900 0.9808678059
#> 25  10.032504 0.0049864808 0.85225192 0.616389456 0.54601463 0.8333116337
#> 26   9.738953 0.2652750881 0.04450751 0.473073209 0.63483123 0.6054639213
#> 27  15.604428 0.7140539121 0.42169618 0.025391043 0.22947027 0.1656253929
#> 28  15.401483 0.8833248129 0.97958003 0.464362063 0.87475770 0.5075860010
#> 29  11.598426 0.0719588706 0.70910878 0.876314595 0.26315656 0.9213615572
#> 30  22.419967 0.6003478721 0.74021944 0.258438713 0.71400168 0.8801359849
#> 31   6.992237 0.0846414506 0.44624366 0.385840196 0.36769365 0.3988182200
#> 32  10.530399 0.1988140659 0.38009132 0.135990085 0.17476498 0.7353523558
#> 33  19.847086 0.5081241641 0.80227878 0.400882215 0.68600768 0.6236524137
#> 34  21.977755 0.7179656278 0.97648304 0.103168434 0.74180842 0.6358334541
#> 35  23.889572 0.9728406055 0.67645554 0.706344246 0.97880851 0.9140141904
#> 36  10.397461 0.8214739454 0.82566148 0.469076067 0.15565227 0.0880800975
#> 37   8.488128 0.0305155648 0.20670491 0.539257291 0.35291808 0.9070779674
#> 38  14.025897 0.3376040715 0.85260841 0.101348271 0.15307282 0.2462114894
#> 39  18.698060 0.6793172057 0.42375187 0.265785902 0.60715259 0.7407385400
#> 40  11.246334 0.2524927731 0.11290691 0.128764630 0.26716735 0.9498405808
#> 41  16.495343 0.9057905758 0.35656974 0.606096197 0.71510168 0.1015809325
#> 42  23.597864 0.8176908845 0.63403191 0.265866829 0.91769619 0.6589987578
#> 43  19.323854 0.4242550391 0.82560730 0.252621680 0.69498959 0.4651354882
#> 44  14.368342 0.0581352876 0.86464668 0.370121471 0.83796469 0.8148452037
#> 45  14.063488 0.4113974599 0.29637662 0.341076544 0.80574213 0.3679854181
#> 46  16.093724 0.7774362189 0.23523198 0.952756745 0.48221441 0.3090288169
#> 47  20.238741 0.7600077491 0.35996427 0.884994609 0.95168397 0.0272900939
#> 48  16.892336 0.1618936695 0.33233874 0.021185879 0.86539555 0.4099761057
#> 49  15.767801 0.7716569337 0.79613133 0.286914685 0.25243823 0.5979735085
#> 50   9.536419 0.1892984733 0.94469213 0.382064596 0.17484849 0.4324351437
#> 51  16.890006 0.8650517790 0.15765001 0.929911307 0.91084710 0.0002038013
#> 52  10.993159 0.0432734324 0.14934288 0.001154730 0.54572437 0.0654351958
#> 53  16.741219 0.4302645724 0.93645664 0.939006025 0.19535360 0.2729003003
#> 54  18.670310 0.4883027766 0.59385577 0.112487184 0.41586598 0.7102491048
#> 55  15.326004 0.1679022135 0.71847951 0.188012942 0.56329559 0.8110345583
#> 56  12.565251 0.2723099228 0.51823811 0.870834145 0.53037101 0.0508010162
#> 57  13.571953 0.5664418542 0.93050028 0.382578627 0.09630942 0.4689941392
#> 58  14.526051 0.9942403871 0.68946617 0.522329734 0.23796757 0.7486345130
#> 59  10.906349 0.3592809127 0.07455763 0.784321454 0.76088597 0.1618195695
#> 60  15.770696 0.7515755724 0.58699456 0.608995861 0.23857511 0.6433369101
#> 61  13.200468 0.9188459744 0.62153336 0.572010342 0.33386413 0.0127919528
#> 62  15.165275 0.4505191611 0.59286612 0.216444159 0.32746945 0.5746817396
#> 63  18.204664 0.5706292351 0.27354225 0.969592683 0.66761848 0.5001979950
#> 64  17.658142 0.3465213720 0.33877773 0.344240971 0.90263934 0.9047999308
#> 65  14.682177 0.6756458520 0.54974357 0.582913484 0.51560775 0.0670099154
#> 66  13.402233 0.6163822697 0.16711914 0.133842469 0.50739050 0.5143311415
#> 67  13.625674 0.7955490339 0.11103459 0.427095284 0.91246536 0.3368847761
#> 68   7.758937 0.5134938993 0.05119809 0.115894507 0.25668613 0.3027861377
#> 69   4.637368 0.0005690118 0.53053807 0.715474398 0.16217414 0.4173761620
#> 70  10.139324 0.0514052720 0.56277804 0.137474032 0.29652110 0.7108454828
#> 71  12.389325 0.8429684825 0.11501862 0.823930671 0.25481224 0.9629575671
#> 72  12.197409 0.2465601282 0.70437974 0.245931110 0.31717505 0.5044448681
#> 73  18.388190 0.3296618462 0.07520944 0.996714146 0.83572332 0.8486035035
#> 74  11.520890 0.5265442270 0.23760233 0.318224024 0.53540244 0.3483987772
#> 75   8.572054 0.1498729351 0.35770824 0.497667168 0.39609426 0.5668962286
#> 76  11.482024 0.1548700077 0.94668445 0.450401472 0.26214785 0.8732300096
#> 77  12.856352 0.6835665777 0.22585451 0.354963427 0.41406614 0.7405214554
#> 78  12.605880 0.2899865448 0.12496412 0.318719039 0.79318805 0.5565725979
#> 79  14.056841 0.8594639292 0.49419164 0.760574816 0.29471374 0.0045515110
#> 80  22.145021 0.9069268992 0.79850353 0.906263766 0.83315352 0.5778649538
#> 81  10.760836 0.6600312220 0.22213483 0.482925944 0.14714913 0.9690910340
#> 82  20.193351 0.3150025082 0.59426908 0.128800116 0.76838131 0.8419886376
#> 83  10.180171 0.7252801226 0.15031994 0.903134433 0.15410268 0.4204075874
#> 84  12.857268 0.1972430279 0.37288395 0.161412123 0.33896239 0.9363293746
#> 85   8.881809 0.0652926126 0.24057065 0.471565024 0.37536625 0.9404777244
#> 86  12.317037 0.3012425804 0.93032003 0.588964753 0.29010223 0.3012445611
#> 87   8.571558 0.5712712964 0.34655776 0.694572859 0.19911126 0.0275595237
#> 88  12.960916 0.1316787691 0.60133943 0.020310450 0.48135941 0.2291809591
#> 89  19.103978 0.5687820101 0.20756611 0.889113101 0.76839987 0.9778400466
#> 90  13.896800 0.1330947741 0.72670317 0.775662492 0.48264546 0.9050342243
#> 91  16.205289 0.2282876193 0.93070729 0.009009782 0.26690255 0.5128316493
#> 92  13.925280 0.4806837339 0.38466815 0.070473357 0.24418923 0.4751092386
#> 93  16.941312 0.9435425298 0.45923936 0.742798108 0.29761078 0.5947223280
#> 94  23.585041 0.7998262872 0.70604558 0.256128479 0.98336860 0.5913213335
#> 95  11.417393 0.3340918350 0.53371406 0.607935532 0.15594806 0.8665356261
#> 96  14.147629 0.1458880980 0.53548490 0.378006948 0.65315099 0.9980839700
#> 97  21.119577 0.4353999950 0.75783154 0.003662417 0.52181626 0.5042093603
#> 98  12.599046 0.3020417586 0.91390889 0.324621239 0.40488362 0.0592789329
#> 99  10.335341 0.2750950810 0.40790019 0.571454250 0.34153083 0.7020047761
#> 100 18.249887 0.8262500311 0.30778956 0.761221548 0.86594112 0.1812446918
#>              x6           x7         x8          x9        x10
#> 1   0.605181459 0.0311939267 0.44965872 0.474430556 0.17301800
#> 2   0.372578762 0.0887080973 0.99694686 0.733113870 0.63027372
#> 3   0.951844210 0.7439887400 0.25880413 0.547647588 0.97529183
#> 4   0.645642431 0.0477283096 0.81951863 0.858631874 0.47926626
#> 5   0.160130299 0.0596022557 0.52106949 0.532094590 0.28024528
#> 6   0.581979284 0.5477562968 0.18488734 0.633268755 0.08717477
#> 7   0.158773502 0.3575467207 0.92165036 0.312998903 0.81400208
#> 8   0.995958803 0.7019681574 0.75113265 0.413347605 0.68433497
#> 9   0.197190441 0.0464429336 0.41304868 0.137553426 0.95527930
#> 10  0.076510635 0.8393243195 0.33290076 0.283656192 0.47750300
#> 11  0.394290680 0.9789608142 0.40082085 0.276244107 0.88409513
#> 12  0.159160965 0.6522488559 0.11830431 0.225956045 0.27646227
#> 13  0.565771605 0.5125643937 0.66709994 0.737031989 0.21885480
#> 14  0.132376131 0.3856457337 0.30236240 0.319441726 0.69527633
#> 15  0.410699643 0.2494750009 0.49005695 0.662589780 0.85850787
#> 16  0.312794350 0.9486472209 0.37666857 0.714928750 0.77000289
#> 17  0.048627757 0.3787378140 0.92405178 0.845063363 0.96209342
#> 18  0.157050259 0.5696307197 0.85633984 0.185738649 0.40764771
#> 19  0.435685530 0.6219776208 0.70175231 0.562042876 0.08046551
#> 20  0.193317769 0.9162687093 0.34299432 0.600176271 0.16952600
#> 21  0.852498348 0.0821403628 0.93766081 0.072483660 0.72117967
#> 22  0.992267430 0.5674253767 0.13896180 0.823995870 0.24160319
#> 23  0.695312970 0.6772589055 0.37378426 0.392794020 0.24449811
#> 24  0.580302832 0.5295008582 0.63379143 0.286064175 0.79545890
#> 25  0.737290514 0.5241456607 0.12785411 0.378323460 0.75962147
#> 26  0.451085591 0.0782050537 0.30449742 0.503644626 0.91019747
#> 27  0.413584440 0.7071137626 0.44859900 0.111460706 0.90858217
#> 28  0.945881257 0.0281042117 0.92248981 0.633271834 0.89347178
#> 29  0.787141449 0.3284668298 0.37709030 0.456755690 0.84228777
#> 30  0.800681989 0.7833664808 0.29794055 0.847923221 0.54371017
#> 31  0.699845273 0.3350773226 0.41442888 0.282253812 0.50112926
#> 32  0.587041139 0.4020150367 0.11787489 0.184324252 0.68590787
#> 33  0.857718518 0.5740697831 0.56397095 0.269160884 0.39397043
#> 34  0.907960838 0.8397472491 0.45881303 0.861369538 0.90115805
#> 35  0.891433122 0.1072088401 0.94075981 0.185943776 0.90444765
#> 36  0.490970159 0.0919665329 0.69400184 0.462276237 0.38179831
#> 37  0.260305499 0.3490059874 0.65991005 0.778763523 0.21369316
#> 38  0.668209213 0.1673681755 0.62726037 0.576612136 0.32394633
#> 39  0.749259676 0.3840076919 0.10810858 0.555713493 0.42300147
#> 40  0.115370568 0.8952052004 0.41636453 0.137661438 0.27289015
#> 41  0.792184972 0.7718279308 0.48631013 0.408459067 0.51037432
#> 42  0.913913373 0.4728496652 0.51036743 0.855744387 0.35335133
#> 43  0.107529560 0.4306397429 0.94496861 0.425768692 0.51772529
#> 44  0.049069265 0.7187078914 0.84792827 0.439417263 0.04000150
#> 45  0.814899431 0.9825139260 0.44902310 0.863252325 0.76488804
#> 46  0.759555861 0.6744440899 0.23348381 0.762450288 0.96982883
#> 47  0.717040052 0.6237124028 0.55543475 0.280952644 0.49259856
#> 48  0.885181972 0.9915004098 0.04680230 0.987688042 0.85611136
#> 49  0.354429267 0.1251676681 0.64307186 0.347842654 0.68280198
#> 50  0.382605365 0.8665508181 0.37843876 0.218780143 0.81434795
#> 51  0.586617316 0.0619861723 0.21621345 0.070207958 0.69921032
#> 52  0.281463075 0.8086633433 0.07786522 0.764758980 0.78502163
#> 53  0.396531604 0.9613011049 0.94262540 0.150003152 0.38973514
#> 54  0.477693542 0.1953248160 0.28340338 0.837943512 0.84303238
#> 55  0.136352813 0.9073695098 0.16199303 0.235072132 0.73080170
#> 56  0.553562591 0.7934069061 0.90383166 0.161564287 0.68641658
#> 57  0.191509547 0.8151038412 0.17097597 0.424137186 0.68119920
#> 58  0.319801679 0.2574579299 0.09973652 0.063093871 0.24827467
#> 59  0.311273515 0.6012528865 0.40586199 0.011670121 0.29276262
#> 60  0.489736654 0.8715508736 0.83962622 0.366843999 0.96650976
#> 61  0.010580119 0.4369509106 0.63812348 0.902484019 0.47271899
#> 62  0.880021336 0.6600062849 0.95273136 0.944477203 0.17811176
#> 63  0.748619123 0.8894117412 0.34247393 0.614323172 0.18281194
#> 64  0.116312593 0.8645935659 0.62734639 0.869429464 0.94208574
#> 65  0.179595005 0.6205700487 0.88543733 0.606220286 0.18762658
#> 66  0.576280995 0.4700410154 0.69751484 0.120708078 0.29253252
#> 67  0.211351240 0.9750166715 0.65603270 0.528520758 0.22245519
#> 68  0.812870071 0.8561156429 0.84540994 0.531322422 0.05215915
#> 69  0.008528329 0.8233643065 0.68669424 0.642060893 0.75653248
#> 70  0.868423827 0.8509351301 0.59933630 0.455682343 0.17605215
#> 71  0.873769116 0.3184813459 0.63138214 0.832083543 0.32986439
#> 72  0.595215278 0.3319889952 0.45442505 0.998439962 0.82127052
#> 73  0.897386551 0.7227046078 0.23091040 0.577683428 0.33350548
#> 74  0.322569185 0.5483409618 0.15259021 0.507942583 0.74123382
#> 75  0.099302771 0.6262102851 0.61149824 0.512107679 0.66015811
#> 76  0.152885180 0.5269158822 0.29757395 0.581978471 0.81284317
#> 77  0.224004767 0.5044637322 0.93052768 0.009639045 0.46563864
#> 78  0.608631474 0.1702083261 0.72714134 0.131025131 0.51623108
#> 79  0.756854297 0.4102603695 0.33373905 0.526454239 0.53860104
#> 80  0.430917728 0.8088284729 0.50952195 0.615642159 0.79611619
#> 81  0.353345245 0.0252314659 0.14586537 0.320850237 0.09055588
#> 82  0.437689098 0.8297493325 0.31944733 0.786132470 0.11387271
#> 83  0.174815357 0.1643891043 0.06778367 0.661611045 0.52499821
#> 84  0.584135804 0.0865170916 0.30470989 0.654257328 0.91446596
#> 85  0.871200818 0.6619940437 0.88528197 0.641658826 0.05146087
#> 86  0.304640589 0.0706656696 0.39910555 0.601995608 0.56509870
#> 87  0.986171405 0.7908917123 0.20083002 0.352823994 0.13495093
#> 88  0.438725859 0.0643991048 0.61970882 0.747814533 0.38802128
#> 89  0.165286306 0.9276355915 0.59946383 0.227795221 0.78316350
#> 90  0.870278344 0.2998997690 0.23028487 0.580738210 0.23000903
#> 91  0.748507656 0.5418102616 0.21522496 0.301186931 0.32406396
#> 92  0.795720276 0.2108449410 0.09606327 0.442556069 0.40255967
#> 93  0.468995362 0.3478310723 0.29872433 0.874471241 0.25975073
#> 94  0.688242224 0.2637047670 0.08132368 0.523058080 0.40147762
#> 95  0.049454222 0.0007590004 0.87957143 0.111005839 0.64907748
#> 96  0.702917031 0.2550472857 0.49797817 0.001788070 0.23905217
#> 97  0.025394215 0.8675307804 0.76169983 0.022281055 0.35801908
#> 98  0.527477048 0.3129886512 0.21769737 0.697959362 0.92573586
#> 99  0.233495052 0.3752802897 0.13103189 0.982251822 0.16365574
#> 100 0.633672248 0.8368462771 0.90448706 0.061223281 0.68831693