Model-Based Variable Importance

Compute model-based variable importance scores for the predictors in a model. (This function is meant for internal use only.)

vi_model(object, ...)

# S3 method for default
vi_model(object, ...)

# S3 method for C5.0
vi_model(object, ...)

# S3 method for constparty
vi_model(object, ...)

# S3 method for earth
vi_model(object, ...)

# S3 method for gbm
vi_model(object, ...)

# S3 method for H2OBinomialModel
vi_model(object, ...)

# S3 method for H2OMultinomialModel
vi_model(object, ...)

# S3 method for H2ORegressionModel
vi_model(object, ...)

# S3 method for lm
vi_model(object, ...)

# S3 method for ml_model_decision_tree_regression
vi_model(object, ...)

# S3 method for ml_model_decision_tree_classification
vi_model(object, ...)

# S3 method for ml_model_gbt_regression
vi_model(object, ...)

# S3 method for ml_model_gbt_classification
vi_model(object, ...)

# S3 method for ml_model_generalized_linear_regression
vi_model(object, ...)

# S3 method for ml_model_linear_regression
vi_model(object, ...)

# S3 method for ml_model_random_forest_regression
vi_model(object, ...)

# S3 method for ml_model_random_forest_classification
vi_model(object, ...)

# S3 method for randomForest
vi_model(object, ...)

# S3 method for RandomForest
vi_model(object, auc = FALSE, ...)

# S3 method for ranger
vi_model(object, ...)

# S3 method for rpart
vi_model(object, ...)

# S3 method for train
vi_model(object, ...)

# S3 method for xgb.Booster
vi_model(object, type = NULL, ...)

Arguments

object
A fitted model object (e.g., a "randomForest" object).

...
Additional optional arguments.

auc
Logical indicating whether or not to compute the AUC-based variable scores described in Janitza et al. (2012). Only available for cforest objects. See varimpAUC for details. Default is FALSE.

type
For xgboost models, the type of variable importance to return. If NULL (the default), "Gain" is used. See xgb.importance for details.

Value

A tidy data frame (i.e., a "tibble" object) with two columns: Variable and Importance. For "lm"/"glm"-like object, an additional column, called Sign, is also included which includes the sign (i.e., POS/NEG) of the original coefficient.

Details

Computes model-based variable importance scores depending on the class of object:

C5.0

Variable importance is measured by determining the percentage of training set samples that fall into all the terminal nodes after the split. For example, the predictor in the first split automatically has an importance measurement of 100 percent since all samples are affected by this split. Other predictors may be used frequently in splits, but if the terminal nodes cover only a handful of training set samples, the importance scores may be close to zero. The same strategy is applied to rule-based models and boosted versions of the model. The underlying function can also return the number of times each predictor was involved in a split by using the option metric = "usage". See C5imp for details.

cforest

Variable importance is measured in a way similar to those computed by importance. Besides the standard version, a conditional version is available that adjusts for correlations between predictor variables. If conditional = TRUE, the importance of each variable is computed by permuting within a grid defined by the predictors that are associated (with 1 - p-value greater than threshold) to the variable of interest. The resulting variable importance score is conditional in the sense of beta coefficients in regression models, but represents the effect of a variable in both main effects and interactions. See Strobl et al. (2008) for details. Note, however, that all random forest results are subject to random variation. Thus, before interpreting the importance ranking, check whether the same ranking is achieved with a different random seed - or otherwise increase the number of trees ntree in ctree_control. Note that in the presence of missings in the predictor variables the procedure described in Hapfelmeier et al. (2012) is performed. See varimp for details.

earth

The variable importance method for earth objects first calculates the decrease in the GCV for each subset relative to the previous subset during earth’s pruning pass. (For multiple response models, GCV's are calculated over all responses.) Then, for each variable, it sums these decreases over all subsets that include the variable. Finally, for ease of interpretation, the summed decreases are scaled so the largest summed decrease is 100. Variables which cause larger net decreases in the GCV are considered more important. See evimp for details.

gbm

Variable importance is computed according to Friedman (2001). For tree-based methods, the approximate relative influence of a variable \(x_j\) is $$\widehat{J}_j^2 = \sum{\text{splits on } x_j}I_t^2$$ where \(I_t\) is the empirical improvement by splitting on \(x_j\) at that point. Friedman's extension to boosted models is to average the relative influence of variable \(x_j\) across all the trees generated by the boosting algorithm. See summary.gbm for details.

H2OModel

See h2o.varimp or visit http://docs.h2o.ai/h2o/latest-stable/h2o-docs/variable-importance.html for details.

lm

In (generalized) linear models, variable importance is based on the absolute value of the corresponding t-statistics. For such models, the sign of the original coefficient is also returned.

ml_feature_importances

The Spark ML library provides standard variable importance for tree-based methods (e.g., random forests). See ml_feature_importances for details.

randomForest

Random forests typically provide two measures of variable importance. The first measure is computed from permuting out-of-bag (OOB) data: for each tree, the prediction error on the OOB portion of the data is recorded (error rate for classification and MSE for regression). Then the same is done after permuting each predictor variable. The difference between the two are then averaged over all trees in the forest, and normalized by the standard deviation of the differences. If the standard deviation of the differences is equal to 0 for a variable, the division is not done (but the average is almost always equal to 0 in that case). The second measure is the total decrease in node impurities from splitting on the variable, averaged over all trees. For classification, the node impurity is measured by the Gini index. For regression, it is measured by residual sum of squares. See importance for details.

cforest

Same approach described in cforest above. See varimp and varimpAUC (if auc = TRUE) for details.

ranger

Variable importance for ranger objects is computed in the usual way for random forests. The approach used depends on the importance argument provided in the initial call to ranger. See importance for details.

rpart

As stated in one of the rpart vignettes. A variable may appear in the tree many times, either as a primary or a surrogate variable. An overall measure of variable importance is the sum of the goodness of split measures for each split for which it was the primary variable, plus "goodness" * (adjusted agreement) for all splits in which it was a surrogate. Imagine two variables which were essentially duplicates of each other; if we did not count surrogates, they would split the importance with neither showing up as strongly as it should. See rpart for details.

train

Various model-specific and model-agnostic approaches that depend on the learning algorithm employed in the original call to train. See varImp for details.

xgboost

For linear models, the variable importance is the absolute magnitude of the estimated coefficients. For that reason, in order to obtain a meaningful ranking by importance for a linear model, the features need to be on the same scale (which you also would want to do when using either L1 or L2 regularization). Otherwise, the approach described in Friedman (2001) for gbms is used. See xgb.importance for details.

Note

Inspired by the varImp function.