Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. See mlbench.friedman1 for details and references.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

Note

This function is mostly used for internal testing.

Examples

gen_friedman()
#>             y          x1         x2         x3         x4         x5
#> 1   15.011217 0.928027559 0.86818910 0.12778479 0.37126514 0.54160008
#> 2   15.617362 0.632786151 0.78043295 0.46678846 0.10962786 0.87279620
#> 3   17.199282 0.370626320 0.28061491 0.19441443 0.85395801 0.73929291
#> 4    7.804093 0.201501012 0.95618046 0.71204565 0.04036009 0.17301965
#> 5   21.418206 0.916043904 0.64435278 0.01208683 0.57979710 0.25568540
#> 6   16.119998 0.870800018 0.98464835 0.84486786 0.63234316 0.60452696
#> 7   15.540104 0.501772149 0.64149951 0.50118417 0.39000222 0.63791025
#> 8   17.894314 0.787037352 0.34793670 0.24299020 0.45448963 0.91534776
#> 9   10.850316 0.133399949 0.34591613 0.66738981 0.42800438 0.92293463
#> 10  15.140936 0.657727109 0.60981555 0.81297528 0.10705957 0.49451297
#> 11  14.051115 0.400631583 0.79744817 0.31367360 0.43628814 0.10328970
#> 12  16.029580 0.772345048 0.40219967 0.67800526 0.47831225 0.47248614
#> 13  17.680811 0.106342572 0.56553444 0.96030009 0.72631563 0.86384756
#> 14  10.103736 0.993839891 0.05850851 0.55920427 0.51215934 0.63322932
#> 15  13.879453 0.887610484 0.91358195 0.34325471 0.38674568 0.78726200
#> 16  14.299118 0.275756894 0.45546538 0.31525782 0.58954279 0.75839077
#> 17  17.424077 0.474025415 0.51850409 0.85717614 0.27973382 0.99780164
#> 18  12.421524 0.059636306 0.38881415 0.88884245 0.48263952 0.79393045
#> 19  15.182100 0.030238627 0.30843007 0.92873990 0.67964156 0.90495598
#> 20  14.113719 0.968499489 0.11325092 0.87740038 0.47675124 0.60646688
#> 21  14.298262 0.529498672 0.23906182 0.60582227 0.65219019 0.73252171
#> 22  13.722351 0.879359917 0.39123990 0.73214967 0.33587752 0.07392884
#> 23  18.600171 0.684657010 0.45290936 0.34870785 0.63775759 0.71316372
#> 24  15.476152 0.671649139 0.28439438 0.14142704 0.22609071 0.99344673
#> 25  12.380672 0.457229698 0.20944055 0.64095440 0.64323263 0.50052659
#> 26  16.899213 0.572100010 0.75079354 0.49317023 0.52130918 0.38114356
#> 27  14.192195 0.574766040 0.44509414 0.31047583 0.41446853 0.42745484
#> 28   2.744995 0.233249805 0.06794720 0.63421243 0.09775226 0.18818207
#> 29  13.561427 0.541338862 0.22667543 0.76811913 0.67072628 0.35510404
#> 30  21.200686 0.477894508 0.57101390 0.14624852 0.89059250 0.44972545
#> 31   7.491041 0.017251855 0.79485996 0.79656682 0.31319037 0.43467045
#> 32   9.886314 0.947369452 0.34071806 0.42107970 0.10670951 0.02794479
#> 33   5.921827 0.084577096 0.58461672 0.34376297 0.25667895 0.24649507
#> 34   9.959102 0.196358106 0.23199268 0.34204696 0.38126270 0.83334334
#> 35  24.217866 0.696352363 0.76760027 0.84022286 0.94130005 0.51640490
#> 36  18.751086 0.308933325 0.89094413 0.22808450 0.83050963 0.28391115
#> 37  16.064916 0.426587059 0.42997629 0.55443234 0.89068156 0.35196549
#> 38  18.654832 0.287835784 0.98425065 0.24438749 0.56629190 0.79183051
#> 39  12.320487 0.516672558 0.48341875 0.65702755 0.14035945 0.65681228
#> 40  14.768520 0.756333916 0.18768922 0.95058792 0.61723732 0.01752624
#> 41  15.008212 0.100261427 0.82012968 0.94862311 0.61970557 0.45520945
#> 42  14.373929 0.586080306 0.83325644 0.48396457 0.29986715 0.25919219
#> 43  19.478474 0.385835596 0.95679158 0.23632482 0.74556099 0.30038002
#> 44  17.035939 0.468633689 0.43303934 0.91842471 0.70981431 0.09838622
#> 45  17.469916 0.980073928 0.65826149 0.95689584 0.34729468 0.14321322
#> 46  22.965072 0.844546806 0.61662452 0.39610285 0.79212095 0.99306096
#> 47  10.728815 0.625345083 0.73655241 0.42343954 0.03501296 0.07521127
#> 48   4.461769 0.671502896 0.06389085 0.49327480 0.07485285 0.46358485
#> 49  14.871803 0.460801910 0.06948297 0.87811625 0.60980295 0.96437558
#> 50   9.956091 0.140419169 0.37831847 0.45229308 0.68508552 0.27372816
#> 51  10.079367 0.277332057 0.50288978 0.55926134 0.08723481 0.99522916
#> 52  16.144715 0.526516733 0.47933950 0.55297529 0.67987999 0.43366900
#> 53  25.997081 0.979514949 0.70255581 0.99917673 0.86725122 0.81731670
#> 54  16.790411 0.991813952 0.55479789 0.46175896 0.27895732 0.80221634
#> 55  15.351997 0.905045445 0.58788148 0.48073698 0.05206537 0.96482198
#> 56  11.551535 0.171565961 0.27221243 0.74999086 0.82421952 0.08581445
#> 57  21.001681 0.265611542 0.88359244 0.85788284 0.88936214 0.56965299
#> 58  12.356677 0.918394280 0.91440857 0.69654237 0.46336045 0.41212335
#> 59  17.626249 0.722241037 0.82413816 0.92643460 0.27726448 0.33538686
#> 60  17.915753 0.378479096 0.53615410 0.90137324 0.53395384 0.72089380
#> 61  11.608102 0.462188396 0.65489714 0.28500307 0.05603563 0.39262885
#> 62  19.707673 0.716373897 0.40628616 0.62291029 0.66585833 0.98448503
#> 63  20.458328 0.766054912 0.43356219 0.54258531 0.93729164 0.51601042
#> 64   8.699384 0.086638817 0.98485921 0.29369212 0.31361313 0.43818175
#> 65  16.736313 0.488905140 0.48100924 0.66427500 0.48320005 0.91855240
#> 66  14.779035 0.032213521 0.55873298 0.16836354 0.85021179 0.68859959
#> 67  13.382734 0.312714351 0.32619606 0.68466414 0.63798300 0.66996790
#> 68  16.439927 0.002584941 0.38412505 0.07287926 0.78810691 0.95790976
#> 69  17.009377 0.197660945 0.88555611 0.64905924 0.90018146 0.48190692
#> 70   6.708631 0.005862298 0.97403861 0.32933181 0.48363088 0.22182491
#> 71   7.909991 0.105120547 0.08317949 0.46213369 0.38396563 0.77235119
#> 72  16.619296 0.779306756 0.66206406 0.92157782 0.04021827 0.54382766
#> 73  12.740852 0.107029717 0.67062342 0.96746911 0.35954754 0.52414767
#> 74  11.380091 0.861860629 0.25992191 0.27909986 0.17731349 0.44104816
#> 75  14.330928 0.681189808 0.09535585 0.19160344 0.65254473 0.79843828
#> 76  11.705664 0.374849067 0.47050856 0.75698823 0.40559883 0.22027266
#> 77  17.981870 0.807005973 0.80382457 0.76807889 0.72249083 0.06515526
#> 78  21.857805 0.501467887 0.67564063 0.03237875 0.51161189 0.71556158
#> 79  23.145815 0.691451428 0.94940063 0.79585309 0.94460735 0.59401146
#> 80  16.871176 0.441635698 0.37671883 0.71114919 0.88412055 0.43896678
#> 81  16.105306 0.971593066 0.91682883 0.17006282 0.78557877 0.54953924
#> 82   8.329304 0.068086573 0.80213278 0.16394792 0.11559714 0.64219482
#> 83  17.248058 0.277309189 0.69577901 0.02396034 0.67237500 0.09112474
#> 84  20.891299 0.788383848 0.51887029 0.82637758 0.68356524 0.49341619
#> 85  16.371147 0.548073788 0.28592805 0.00728124 0.18895710 0.95774968
#> 86  23.904320 0.439798625 0.99862824 0.10015306 0.69527295 0.79703914
#> 87   9.968615 0.582232314 0.16302308 0.24477611 0.37240254 0.45208699
#> 88  20.254312 0.912940030 0.76279186 0.38971042 0.83693223 0.67600502
#> 89   6.007225 0.216799448 0.02445868 0.36730736 0.18131958 0.73402940
#> 90  15.107182 0.508444319 0.23201441 0.29545443 0.87311741 0.37963744
#> 91  22.188672 0.935941602 0.84196227 0.98760562 0.97128141 0.32921006
#> 92  21.333290 0.930981670 0.44784109 0.44690295 0.69428288 0.92050228
#> 93  15.761905 0.078153792 0.20959445 0.03263834 0.94473749 0.31606266
#> 94  15.369993 0.661558503 0.75498099 0.73194192 0.30802347 0.24656567
#> 95  24.773027 0.504792338 0.72311510 0.01646533 0.63923236 0.92518609
#> 96   6.429644 0.079875728 0.01971191 0.21424688 0.02961796 0.89266254
#> 97   4.238757 0.002086343 0.95793856 0.58938934 0.01201998 0.79237289
#> 98  16.365694 0.540948429 0.77982982 0.67401817 0.11191973 0.97818299
#> 99  13.126630 0.157224735 0.40781335 0.03304105 0.59895507 0.12795916
#> 100 15.495146 0.431135125 0.73094312 0.36421439 0.27544296 0.81745347
#>             x6          x7         x8           x9         x10
#> 1   0.66653145 0.550015410 0.76766640 0.9650785702 0.978117280
#> 2   0.52387645 0.109289986 0.28783645 0.5587809035 0.467956529
#> 3   0.83247081 0.482667100 0.10508655 0.0026304119 0.398290447
#> 4   0.42774729 0.690652648 0.63019161 0.3712419253 0.502596827
#> 5   0.02406933 0.750984840 0.77476928 0.6818092742 0.477677812
#> 6   0.48002291 0.875439102 0.67833387 0.0859072856 0.723185581
#> 7   0.54420622 0.856983725 0.99984863 0.9591735939 0.986509347
#> 8   0.30184299 0.124779644 0.24697301 0.3756451020 0.607789966
#> 9   0.18491759 0.235410240 0.77243877 0.9552118548 0.348995293
#> 10  0.37011362 0.903017866 0.64173216 0.0946726704 0.410259061
#> 11  0.15520864 0.436727899 0.91325843 0.7531174992 0.270249070
#> 12  0.72580448 0.631451685 0.73362766 0.5961748285 0.884834194
#> 13  0.38711210 0.685201029 0.03314948 0.1477832196 0.503985950
#> 14  0.35216916 0.062761624 0.83448169 0.6963498048 0.701912774
#> 15  0.99301777 0.028264773 0.05017491 0.3420383276 0.337836640
#> 16  0.85639693 0.852932025 0.65943429 0.2466902647 0.324442123
#> 17  0.34837392 0.778535618 0.25909856 0.9581220152 0.983576211
#> 18  0.99183043 0.037995263 0.42812983 0.2072116183 0.496675399
#> 19  0.86466355 0.683828282 0.66398431 0.9802682598 0.070524038
#> 20  0.21861263 0.693640037 0.24547323 0.5085619763 0.970124071
#> 21  0.87290733 0.220077557 0.65057534 0.3759201681 0.235627398
#> 22  0.57695613 0.469844026 0.71848470 0.0486618015 0.223642083
#> 23  0.28255106 0.923913826 0.49560666 0.2968211703 0.124602350
#> 24  0.09146729 0.767172322 0.53090338 0.6874899382 0.169910871
#> 25  0.17736456 0.456660097 0.87662332 0.3377936454 0.127875301
#> 26  0.47562988 0.761600550 0.43078692 0.5493953177 0.639769726
#> 27  0.38710091 0.752808762 0.87777997 0.8074811040 0.210232693
#> 28  0.07365412 0.491128261 0.91691338 0.7400463219 0.342927222
#> 29  0.87382203 0.914959561 0.80538165 0.4218737460 0.993826717
#> 30  0.09838254 0.878033062 0.22153729 0.7465715150 0.522418012
#> 31  0.96580435 0.780594654 0.03252142 0.0170752136 0.894790446
#> 32  0.97669054 0.885739977 0.76756146 0.2801137855 0.159248836
#> 33  0.68297299 0.725453006 0.67791248 0.7146302699 0.067293887
#> 34  0.85781335 0.868947522 0.74575029 0.2180433657 0.723391416
#> 35  0.52468723 0.239178060 0.19545495 0.3426197150 0.218654526
#> 36  0.28688805 0.640270370 0.52947622 0.4860946555 0.597464886
#> 37  0.74438635 0.921434219 0.86361339 0.2956594473 0.124660186
#> 38  0.62224371 0.273144853 0.03435079 0.8688688970 0.180205426
#> 39  0.97582986 0.539830090 0.12613023 0.7080726644 0.839665562
#> 40  0.94017284 0.909498676 0.39148929 0.2317488412 0.807895935
#> 41  0.21488036 0.471091736 0.78029207 0.7893189187 0.339445215
#> 42  0.63399320 0.249009586 0.15922846 0.6016295548 0.601717093
#> 43  0.29175457 0.515118479 0.72395483 0.2173144864 0.256999088
#> 44  0.58658586 0.445304808 0.11837270 0.1882908070 0.097694466
#> 45  0.12885259 0.861717390 0.44013645 0.7303674868 0.558897309
#> 46  0.39829348 0.441197833 0.25440320 0.8846553627 0.376339515
#> 47  0.36320548 0.911988780 0.27200908 0.7479972423 0.683729710
#> 48  0.51591539 0.009186007 0.59597918 0.5562674277 0.117525956
#> 49  0.21632537 0.445051267 0.57289347 0.0717247135 0.355477426
#> 50  0.37682732 0.261930517 0.07532414 0.0351810509 0.852195424
#> 51  0.28070321 0.865729444 0.49610853 0.0496929921 0.009848715
#> 52  0.54089192 0.465138446 0.42056673 0.2043203793 0.608554971
#> 53  0.04871538 0.387150318 0.59973788 0.2121785304 0.120526727
#> 54  0.10392306 0.892417365 0.04953585 0.7616382262 0.045770801
#> 55  0.77438982 0.255271181 0.24908242 0.7481696780 0.934910491
#> 56  0.67482756 0.662540321 0.19961701 0.6307045645 0.421319699
#> 57  0.02657235 0.152472844 0.21237341 0.8518509718 0.292108134
#> 58  0.15364551 0.041097763 0.67097056 0.7620710412 0.173811029
#> 59  0.42272721 0.326579318 0.52631656 0.1620942338 0.988964739
#> 60  0.54698992 0.573182128 0.82139899 0.2296430122 0.379289672
#> 61  0.97115919 0.856403784 0.27267502 0.0330121680 0.124106386
#> 62  0.84867392 0.424555891 0.76325505 0.8664087972 0.446321585
#> 63  0.98192967 0.746969166 0.89396226 0.2821210856 0.371556330
#> 64  0.72915852 0.636969819 0.85983604 0.1043299169 0.915742398
#> 65  0.98441934 0.838804665 0.22639941 0.6998189061 0.393983506
#> 66  0.82599302 0.500919412 0.95483553 0.6506008462 0.176385239
#> 67  0.74345612 0.883136018 0.35103400 0.4079634142 0.250502597
#> 68  0.67537699 0.720257379 0.35120966 0.2076517732 0.814686354
#> 69  0.94247517 0.169301706 0.76188172 0.7838959824 0.370211543
#> 70  0.29671047 0.347173683 0.58847983 0.5876605695 0.931645184
#> 71  0.70227069 0.323358465 0.16493367 0.2705651452 0.760856272
#> 72  0.91271488 0.522552675 0.93470262 0.9763128029 0.170257425
#> 73  0.60776491 0.195973985 0.80357658 0.5277703388 0.585484079
#> 74  0.08484402 0.622975350 0.72049649 0.5019180186 0.303493992
#> 75  0.42027553 0.673768653 0.27971170 0.5393672534 0.123312983
#> 76  0.45107443 0.146693246 0.68820855 0.5804427962 0.925233021
#> 77  0.53191233 0.814188991 0.58505368 0.3886140115 0.222452674
#> 78  0.13845940 0.600851081 0.82373428 0.7392512404 0.189938827
#> 79  0.24915372 0.636284345 0.62754211 0.0057790936 0.310507899
#> 80  0.38730836 0.232025593 0.54897299 0.9339925982 0.677229591
#> 81  0.60348792 0.582359949 0.97701678 0.5044764206 0.360172970
#> 82  0.80320427 0.649527824 0.17371482 0.1802249348 0.927258149
#> 83  0.31054726 0.473346722 0.47905547 0.2686649794 0.410745379
#> 84  0.21534486 0.521855675 0.67115081 0.2856403359 0.746765224
#> 85  0.21212699 0.962416184 0.34229264 0.4530142238 0.117128020
#> 86  0.59922707 0.969441712 0.58506156 0.5629261434 0.291638010
#> 87  0.12086655 0.152528613 0.86820448 0.9473500950 0.567477844
#> 88  0.35700736 0.548785103 0.24548555 0.1686221259 0.788605310
#> 89  0.46992428 0.322719622 0.31959777 0.0002908951 0.915483615
#> 90  0.48924495 0.337137345 0.79090592 0.0017725758 0.463239050
#> 91  0.08917308 0.320080395 0.33547448 0.8257024728 0.275465864
#> 92  0.36720251 0.693340738 0.02543847 0.3237691270 0.407288683
#> 93  0.25450057 0.834965316 0.49670450 0.0260559008 0.462093781
#> 94  0.52116721 0.317677672 0.40114180 0.7262006637 0.899225596
#> 95  0.12296857 0.657580597 0.93241442 0.7835480885 0.021150852
#> 96  0.68118458 0.743931412 0.84334804 0.6704279722 0.260274333
#> 97  0.10983796 0.605542281 0.21598949 0.7793204396 0.696119128
#> 98  0.75025096 0.804082083 0.92993483 0.3367602522 0.738093848
#> 99  0.63139774 0.546061827 0.54952469 0.5521352317 0.526687683
#> 100 0.96247233 0.964195222 0.85828676 0.5118112578 0.084802875