Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. These data were originally described in Friedman (1991) and Breiman (1996). For details, see sklearn.datasets.make_friedman1.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

References

Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The Annals of Statistics 19 (1), pages 1-67.

Examples

gen_friedman()
#>             y         x1          x2         x3         x4          x5
#> 1   16.004258 0.48681190 0.854560890 0.46137475 0.41632914 0.425863413
#> 2   17.190616 0.63850214 0.233965138 0.03931026 0.78069661 0.147350063
#> 3   16.617573 0.42693146 0.835591768 0.72493013 0.28027390 0.800559559
#> 4   19.597918 0.69141305 0.373826457 0.27750798 0.68313824 0.867213433
#> 5   21.810275 0.77797011 0.780834123 0.28079918 0.81597210 0.636517575
#> 6   10.729109 0.30812186 0.705781814 0.73597980 0.33260271 0.004782590
#> 7   14.320435 0.78631847 0.348107484 0.09237952 0.27938738 0.133421473
#> 8   19.535508 0.60691072 0.801859443 0.69231001 0.49880564 0.759636998
#> 9   15.744591 0.85907093 0.783380867 0.46061161 0.34984954 0.781685917
#> 10  21.416105 0.82211062 0.310290070 0.58072621 0.95592514 0.916371646
#> 11  10.706178 0.28541652 0.542996578 0.29795667 0.27815461 0.523505823
#> 12   8.063473 0.38136799 0.505088628 0.24386766 0.03669076 0.117748061
#> 13  13.260445 0.73219669 0.103538398 0.63760662 0.95451069 0.170087027
#> 14  18.582963 0.66405776 0.388559964 0.48741455 0.94279894 0.398755051
#> 15  13.067325 0.34552494 0.702196913 0.62684845 0.36805176 0.434461756
#> 16  13.765782 0.04427833 0.924009398 0.80929019 0.80629731 0.487343654
#> 17  13.619529 0.28669630 0.246152389 0.28744338 0.79586497 0.505168315
#> 18  10.606604 0.24916725 0.232852984 0.16241165 0.20751229 0.905977423
#> 19  16.788107 0.53289045 0.759034161 0.16392721 0.46927897 0.056512806
#> 20  17.038666 0.94263312 0.652660663 0.89425627 0.14174522 0.639012057
#> 21  13.625781 0.01934215 0.827592164 0.08857593 0.89725043 0.142949865
#> 22  14.672896 0.99534765 0.279921843 0.87679166 0.07407103 0.713112807
#> 23  15.364084 0.67900649 0.306089681 0.94785282 0.41213039 0.254710316
#> 24   9.498943 0.72746925 0.141634461 0.29557881 0.37763565 0.346040505
#> 25   9.949699 0.05313648 0.853700958 0.56972134 0.81953816 0.026013865
#> 26  19.189087 0.75831112 0.213609558 0.25978785 0.93279118 0.780463972
#> 27  20.681957 0.70638865 0.620064068 0.24600348 0.87093888 0.172627212
#> 28  18.288390 0.95586155 0.619915097 0.22775554 0.37970779 0.663977504
#> 29  14.036542 0.34152542 0.682499397 0.84601312 0.30030786 0.377479340
#> 30   8.253271 0.22939649 0.963768314 0.61383438 0.01567223 0.300973915
#> 31  10.589563 0.67301747 0.368319741 0.67303883 0.10931868 0.350964558
#> 32   9.537200 0.09928076 0.297129260 0.01493150 0.21086985 0.385667540
#> 33  16.454707 0.85069775 0.271640491 0.79922452 0.71524543 0.150912791
#> 34   8.526208 0.66204358 0.137757749 0.66220868 0.33902436 0.355765326
#> 35   9.426022 0.83857548 0.045895619 0.76632657 0.50639927 0.352611043
#> 36  15.128284 0.65788463 0.335906105 0.29511898 0.34940403 0.884454354
#> 37  19.400211 0.58419523 0.957532233 0.39979602 0.91499682 0.043013364
#> 38  10.763657 0.17537294 0.257989588 0.56217637 0.75396917 0.379090654
#> 39  24.325268 0.66133427 0.869326131 0.20023315 0.92729730 0.746833655
#> 40  20.092996 0.52798573 0.957021768 0.16522612 0.43153201 0.681802785
#> 41  23.558681 0.86752483 0.404717089 0.75886603 0.97676822 0.710866165
#> 42  19.069815 0.78318547 0.254892744 0.09017591 0.75062910 0.469135608
#> 43   7.634177 0.03320663 0.121329135 0.34045463 0.50102869 0.399724711
#> 44  13.152926 0.98949290 0.148240795 0.82784096 0.34666070 0.610139551
#> 45   7.373990 0.14975105 0.423030155 0.27443060 0.10902323 0.674438052
#> 46  11.557565 0.20449880 0.498953031 0.20794405 0.64296013 0.049324810
#> 47  23.472400 0.75172972 0.843373979 0.36905288 0.94655935 0.895236755
#> 48   9.988036 0.91908646 0.068247122 0.71704253 0.45141070 0.540919727
#> 49  18.286847 0.66505137 0.585394813 0.20183357 0.54700572 0.285610262
#> 50  14.090384 0.56792096 0.017736985 0.06992703 0.57095054 0.883247489
#> 51  10.751894 0.64164881 0.364722385 0.12274733 0.05016204 0.134332892
#> 52  14.332590 0.37355189 0.371583528 0.37284361 0.64480542 0.641885565
#> 53  19.770781 0.90755256 0.650434283 0.51753396 0.74269025 0.532572363
#> 54  13.122339 0.34906802 0.659298811 0.14701448 0.18414778 0.438645379
#> 55  19.699753 0.83658483 0.412022420 0.29074831 0.93361809 0.124604392
#> 56  12.278088 0.48628757 0.117003911 0.59959033 0.82878236 0.424732280
#> 57  11.615436 0.25793487 0.463965650 0.67942219 0.51506542 0.464081346
#> 58  21.680211 0.98234430 0.559486266 0.26758683 0.75199872 0.672870615
#> 59  11.268061 0.71658398 0.339364589 0.30112174 0.25316737 0.212657275
#> 60  11.560335 0.83577993 0.166738186 0.47930017 0.60732563 0.227992547
#> 61  10.589703 0.84068295 0.003786570 0.63194295 0.55550598 0.913280469
#> 62  19.734057 0.94152475 0.977840924 0.93408891 0.92542532 0.857124415
#> 63  10.993307 0.84366409 0.679824525 0.55177244 0.09999405 0.031652359
#> 64  10.902276 0.01790754 0.439787539 0.47887470 0.94695896 0.230908046
#> 65  16.808997 0.93350571 0.314297716 0.66282196 0.53804799 0.591017207
#> 66  10.687664 0.18811810 0.077688647 0.46482363 0.62813184 0.799680480
#> 67  15.571169 0.80576590 0.881355740 0.22264569 0.41280780 0.381758855
#> 68  18.137624 0.81729285 0.287769772 0.12226053 0.74416377 0.225910783
#> 69  18.956850 0.68254790 0.587704190 0.03340033 0.39059331 0.217378195
#> 70  18.522635 0.97269288 0.235262827 0.18774756 0.92441481 0.156608932
#> 71  17.500236 0.85400062 0.897141295 0.32568579 0.99235506 0.078432295
#> 72  11.757172 0.45645770 0.342318100 0.87220318 0.37319174 0.061701847
#> 73  18.793313 0.82018978 0.939842280 0.79908875 0.55275477 0.940274754
#> 74  12.758358 0.63155440 0.202082169 0.02520818 0.17295903 0.532553327
#> 75  19.281390 0.25067659 0.853603688 0.76451970 0.69200940 0.952136923
#> 76  15.488551 0.26540909 0.588013205 0.86463905 0.39155581 0.839863650
#> 77  16.600683 0.96420943 0.764233994 0.78274457 0.46913390 0.563972442
#> 78  14.713062 0.22609516 0.493023776 0.26881551 0.53755426 0.955716009
#> 79  11.977472 0.23805146 0.182082188 0.17300284 0.65882938 0.356307692
#> 80  10.025654 0.89473860 0.999525146 0.14950306 0.40806456 0.056172378
#> 81  18.488083 0.55796304 0.710131911 0.89475588 0.41821521 0.387330521
#> 82  15.051802 0.47683006 0.915772210 0.88826356 0.17524390 0.089917012
#> 83   6.168537 0.91515464 0.093801769 0.59313256 0.31547053 0.003703875
#> 84  16.198437 0.50197664 0.068456021 0.61271223 0.99658203 0.965599259
#> 85  14.666017 0.29934019 0.122394820 0.70068596 0.98144770 0.583912069
#> 86   7.078456 0.12602068 0.320340649 0.28392566 0.41674902 0.153703178
#> 87  11.294797 0.02328787 0.892295834 0.29069625 0.60229628 0.764272182
#> 88  11.640738 0.40392790 0.287106070 0.03408497 0.19393290 0.374494878
#> 89  13.182910 0.67475230 0.363371446 0.56688957 0.41891229 0.393806637
#> 90  14.927634 0.50266743 0.280262201 0.72403062 0.73632455 0.446247931
#> 91  10.740506 0.07943226 0.477118348 0.34715285 0.87346708 0.060941179
#> 92  15.254052 0.76633929 0.167160649 0.67845543 0.66928315 0.758281781
#> 93  10.170362 0.15380592 0.318547365 0.92769605 0.18558102 0.603362613
#> 94  18.628545 0.23324016 0.817493711 0.11999019 0.78562236 0.482028900
#> 95  11.729214 0.59778737 0.007275813 0.94318157 0.49882768 0.547365508
#> 96  24.802895 0.84085391 0.514042865 0.88941496 0.74665649 0.901769097
#> 97  14.598534 0.08641988 0.976045435 0.80826481 0.58897011 0.824543996
#> 98  15.817543 0.60706769 0.901173368 0.83595324 0.07070052 0.612242390
#> 99   9.924087 0.01186816 0.192172108 0.75185388 0.47169970 0.792779175
#> 100 12.769951 0.28917309 0.232875240 0.40986467 0.81068021 0.460743416
#>              x6          x7          x8          x9        x10
#> 1   0.060940081 0.218619842 0.459664619 0.947586665 0.20663897
#> 2   0.680445127 0.881185908 0.573548796 0.455810402 0.47406560
#> 3   0.803962911 0.748578238 0.403673562 0.444485249 0.40527456
#> 4   0.330934667 0.612023658 0.613945051 0.828754810 0.41372006
#> 5   0.836526289 0.791179617 0.767778325 0.560495309 0.67772193
#> 6   0.105866324 0.529143142 0.003227110 0.727505995 0.99487431
#> 7   0.239136878 0.511157620 0.463155464 0.939445548 0.20753867
#> 8   0.761442949 0.061624395 0.478592061 0.804514406 0.90205689
#> 9   0.985014238 0.602984187 0.811805757 0.705605592 0.07226518
#> 10  0.731627790 0.547401533 0.511561444 0.655437261 0.09095566
#> 11  0.996215126 0.120773211 0.968544949 0.310046385 0.18077775
#> 12  0.943059926 0.625854586 0.451553049 0.928418229 0.07507598
#> 13  0.111095259 0.484418190 0.733188885 0.459188519 0.93936006
#> 14  0.031787673 0.209827383 0.701901909 0.688930298 0.77396049
#> 15  0.308491633 0.353743039 0.756843362 0.643296141 0.09653514
#> 16  0.644992244 0.621993710 0.067071428 0.661989347 0.72162516
#> 17  0.015900322 0.681852613 0.455959291 0.480765560 0.52762746
#> 18  0.312353626 0.009935961 0.493619153 0.806100092 0.84917015
#> 19  0.331310340 0.784542835 0.054218127 0.399089943 0.86993264
#> 20  0.143738320 0.514003187 0.482053395 0.500799541 0.78242142
#> 21  0.689244658 0.294256262 0.822894737 0.669721918 0.56027131
#> 22  0.143564009 0.192108605 0.342938086 0.580366435 0.51928929
#> 23  0.124880360 0.161287538 0.037619785 0.729642778 0.43846243
#> 24  0.152271881 0.210781540 0.850797251 0.855638933 0.44896396
#> 25  0.735151459 0.570147376 0.716051252 0.387275586 0.70736157
#> 26  0.191510390 0.010619027 0.192964875 0.506600111 0.54101489
#> 27  0.895277515 0.462958924 0.747208879 0.440723592 0.44297880
#> 28  0.442210175 0.316611744 0.181750269 0.794021430 0.02524313
#> 29  0.725990596 0.761538408 0.150551342 0.288950684 0.97844911
#> 30  0.592035108 0.169682043 0.149329605 0.642700075 0.67520989
#> 31  0.313103905 0.413818424 0.571162875 0.895917939 0.31619599
#> 32  0.099071316 0.132128711 0.554896752 0.712193127 0.77224176
#> 33  0.761797328 0.848752375 0.415704421 0.604048606 0.51989640
#> 34  0.514539573 0.103947730 0.643566482 0.133184023 0.89896707
#> 35  0.126993078 0.204644089 0.589278607 0.014696075 0.73496029
#> 36  0.480614223 0.142012031 0.502939152 0.659132935 0.12116264
#> 37  0.529881594 0.701897858 0.772195464 0.385836676 0.05442022
#> 38  0.512287199 0.228002047 0.777741790 0.640518297 0.37322869
#> 39  0.818204291 0.597731920 0.898624761 0.544441006 0.48036606
#> 40  0.263570308 0.058304883 0.185875483 0.020852469 0.79709812
#> 41  0.575239548 0.085831387 0.003649816 0.085122694 0.55699828
#> 42  0.043911399 0.703081927 0.404914220 0.346472173 0.14575330
#> 43  0.563091956 0.275666144 0.465727412 0.638893410 0.91525944
#> 44  0.056795009 0.625039192 0.114429964 0.763661725 0.64739827
#> 45  0.869537722 0.481230214 0.660998754 0.584850597 0.35015185
#> 46  0.169645325 0.162083877 0.113590301 0.853298074 0.98551212
#> 47  0.654906135 0.623367399 0.296049605 0.184455995 0.84461505
#> 48  0.813826193 0.641873605 0.123395729 0.244064771 0.18138980
#> 49  0.006665684 0.444909158 0.223069615 0.254802058 0.94960459
#> 50  0.280707166 0.645511346 0.709198708 0.878724405 0.65595549
#> 51  0.449564583 0.066077120 0.835829101 0.847710915 0.81758008
#> 52  0.820717620 0.789126059 0.268236581 0.736481237 0.77400708
#> 53  0.611357286 0.556894590 0.589966203 0.851852106 0.10523739
#> 54  0.314421845 0.988239036 0.365324286 0.214313712 0.81067618
#> 55  0.504150994 0.008326067 0.979020421 0.520721707 0.90963442
#> 56  0.170635836 0.031910183 0.641698274 0.619547703 0.22223670
#> 57  0.153934190 0.251441214 0.977109678 0.746170353 0.90143123
#> 58  0.748798251 0.424600797 0.058784112 0.963038840 0.44663062
#> 59  0.800845830 0.050406727 0.011864729 0.106421710 0.62974473
#> 60  0.521239672 0.580963297 0.080027009 0.547273749 0.66139832
#> 61  0.965652359 0.524809101 0.828726630 0.128897642 0.21656126
#> 62  0.585446123 0.170050133 0.945188337 0.422906543 0.12767548
#> 63  0.374685861 0.971438063 0.005105530 0.656417205 0.24430470
#> 64  0.986206700 0.036075082 0.204774484 0.467459401 0.81470334
#> 65  0.029265686 0.463792012 0.643861896 0.538708197 0.15761692
#> 66  0.578512490 0.672093972 0.855779017 0.081111051 0.68974210
#> 67  0.277229002 0.001566010 0.417831972 0.423333576 0.48618521
#> 68  0.296267914 0.038521656 0.665277467 0.381922779 0.08733217
#> 69  0.062232896 0.358096567 0.805208212 0.087945131 0.26081447
#> 70  0.507230302 0.445598125 0.129567425 0.239824070 0.61515989
#> 71  0.883192962 0.898561200 0.225106215 0.381404544 0.66491042
#> 72  0.445139494 0.798542147 0.631360549 0.763079448 0.37930803
#> 73  0.338407275 0.334779249 0.710675666 0.656298424 0.04474986
#> 74  0.144883063 0.351377906 0.913218759 0.653570792 0.54383016
#> 75  0.287825262 0.496957100 0.619113757 0.591041674 0.28150960
#> 76  0.151020752 0.639492054 0.550524612 0.412430681 0.32582163
#> 77  0.271489745 0.890239056 0.794838918 0.398868840 0.43415918
#> 78  0.939219849 0.147719556 0.915753975 0.266160668 0.61421796
#> 79  0.483586594 0.706494146 0.483055972 0.576164000 0.81376446
#> 80  0.952745022 0.393328290 0.369904404 0.702081015 0.71554226
#> 81  0.788842335 0.278614218 0.517253958 0.991534388 0.53967087
#> 82  0.581547714 0.144217548 0.680271830 0.282528550 0.97704681
#> 83  0.037537344 0.553284194 0.580729727 0.439574573 0.26128947
#> 84  0.562256702 0.390895223 0.253185890 0.931511228 0.95349208
#> 85  0.181867038 0.017102353 0.063916326 0.093837686 0.76563168
#> 86  0.570067758 0.223020279 0.892349249 0.785752190 0.45662483
#> 87  0.941815388 0.694731108 0.045712603 0.697217774 0.15117451
#> 88  0.519291770 0.855090562 0.190505332 0.243971932 0.24836411
#> 89  0.618373664 0.515685107 0.412341356 0.938209138 0.01039147
#> 90  0.765881778 0.599906300 0.613791870 0.493707179 0.60006984
#> 91  0.992043673 0.748782483 0.556992820 0.119921436 0.10638014
#> 92  0.094131473 0.181056903 0.555123862 0.984366343 0.78282070
#> 93  0.966871710 0.741605376 0.053775917 0.324312976 0.26912155
#> 94  0.990510182 0.599045724 0.980053324 0.189629599 0.33666098
#> 95  0.574063676 0.971672369 0.992454848 0.176421642 0.17153048
#> 96  0.060912089 0.806568764 0.364127242 0.002105691 0.72942866
#> 97  0.363061381 0.677878273 0.788224100 0.957070542 0.37728700
#> 98  0.435520820 0.805944814 0.680775681 0.325668049 0.53431836
#> 99  0.707710207 0.983456140 0.157901129 0.058567173 0.22902009
#> 100 0.167980285 0.486589333 0.572332644 0.155346128 0.33518173