Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. These data were originally described in Friedman (1991) and Breiman (1996). For details, see sklearn.datasets.make_friedman1.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

References

Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The Annals of Statistics 19 (1), pages 1-67.

Examples

gen_friedman()
#>             y           x1           x2         x3         x4          x5
#> 1   23.840987 0.6759721697 0.5125813521 0.07160379 0.99978627 0.271178257
#> 2    9.623661 0.9856958392 0.9642838873 0.03351095 0.24955093 0.250500011
#> 3   22.654672 0.4825402873 0.7847610181 0.13789961 0.69669849 0.761911532
#> 4   20.232250 0.4456876006 0.9189490753 0.30500337 0.54824237 0.847765672
#> 5    9.185394 0.0006588304 0.9077979270 0.79091867 0.65420285 0.172688548
#> 6    9.250396 0.9212993807 0.0260542023 0.64030548 0.36272326 0.896859805
#> 7   22.789935 0.5765405579 0.4894680525 0.19219761 0.88930341 0.850170494
#> 8   12.169586 0.6954413948 0.3028189589 0.13943872 0.13401661 0.425986421
#> 9   19.340471 0.6217532922 0.7752467443 0.23420838 0.36659762 0.806701331
#> 10   3.793405 0.8850848931 0.1239252316 0.59525327 0.02236348 0.021312950
#> 11  19.413926 0.4397395041 0.3611886650 0.39094931 0.97708755 0.904544010
#> 12   6.498691 0.2040736177 0.0627803814 0.68211590 0.31129201 0.417710888
#> 13  20.030404 0.8196485452 0.8272371439 0.51922226 0.72653047 0.839141774
#> 14  24.325077 0.5386889591 0.8067489006 0.77988461 0.89965647 0.802761870
#> 15   6.840599 0.0492286240 0.6635464525 0.59434994 0.26024182 0.630479756
#> 16   4.467399 0.3394161433 0.1120156294 0.44474067 0.31079865 0.023790262
#> 17  15.773778 0.6430535160 0.3247046669 0.91971075 0.16999976 0.899287550
#> 18  21.327641 0.9076674583 0.7251544110 0.14231176 0.71817093 0.560615089
#> 19  13.627077 0.5539126799 0.0861778562 0.37922109 0.71879158 0.930752811
#> 20   5.346005 0.0228288048 0.3503480782 0.58003563 0.32711820 0.346533943
#> 21  11.908613 0.0734738058 0.7593257639 0.85673325 0.64319400 0.238326720
#> 22  12.656367 0.8010455330 0.7365703774 0.48800525 0.23216312 0.115233585
#> 23   9.119648 0.5428822343 0.0767615237 0.16120203 0.28123957 0.558483608
#> 24  15.670994 0.5024930094 0.8700824017 0.21632527 0.18057260 0.481429608
#> 25   8.218650 0.0990956256 0.6461682799 0.79556184 0.41996217 0.061638433
#> 26   9.771555 0.8087382587 0.1461430411 0.84577016 0.38154611 0.005737968
#> 27  17.872139 0.0286039875 0.8538238918 0.99068950 0.93912513 0.585702214
#> 28  23.455408 0.8855307666 0.8788125345 0.94815567 0.80214038 0.996345679
#> 29  10.573043 0.6016014367 0.3496340397 0.27380223 0.10963093 0.441241066
#> 30   7.177664 0.2972732172 0.4656577776 0.75771528 0.05869202 0.206455562
#> 31  24.726120 0.5657286951 0.6616219610 0.99093952 0.80541971 0.549726494
#> 32  11.051004 0.7170329434 0.7876186040 0.30978760 0.01633929 0.061329732
#> 33   9.372561 0.4290181776 0.4004281636 0.70406353 0.12615941 0.444348910
#> 34  10.023614 0.1576340361 0.6216970964 0.78210182 0.06048809 0.947721819
#> 35  11.584755 0.4258746603 0.0042340204 0.03112690 0.53898954 0.351290324
#> 36  16.369986 0.3514485436 0.8470020872 0.39517534 0.53379761 0.563543642
#> 37  11.801550 0.2609182710 0.4988422689 0.53310193 0.41148769 0.722206617
#> 38   6.551574 0.8293840422 0.0536368384 0.33471232 0.14434654 0.669263439
#> 39   5.595860 0.2120036017 0.3893485165 0.65326158 0.23088470 0.062985555
#> 40  20.208615 0.3250550434 0.6858599265 0.91424940 0.60875733 0.879544277
#> 41  16.986450 0.6789210238 0.4106222112 0.45347471 0.80048570 0.211604108
#> 42  13.495114 0.6430051320 0.2600766481 0.92377445 0.09132057 0.781262683
#> 43  17.035465 0.4125444102 0.5177675276 0.98777078 0.46859677 0.309003265
#> 44  12.326859 0.1176328119 0.1946966413 0.62589850 0.68688279 0.884768495
#> 45  15.761771 0.6594399249 0.6072569215 0.23213414 0.43492972 0.110364238
#> 46  23.163308 0.9416496824 0.4092349561 0.99005480 0.82986953 0.113867485
#> 47  11.205558 0.9996683616 0.0294578678 0.53462244 0.88237138 0.299841770
#> 48  16.204456 0.5281595406 0.0010634891 0.81008305 0.97356747 0.924107481
#> 49  20.309441 0.6546782369 0.9711225589 0.86645495 0.64830269 0.427723981
#> 50  11.714903 0.5696514260 0.8825622501 0.29380554 0.05614968 0.047195012
#> 51  16.835248 0.9053750585 0.8535252616 0.20203872 0.73011502 0.242909296
#> 52  15.972953 0.9146019304 0.5891615695 0.50853111 0.44354240 0.356688302
#> 53  12.931350 0.3218688134 0.7612053347 0.73249181 0.45506004 0.070664430
#> 54  11.437905 0.3383209873 0.5550955092 0.45884185 0.45195129 0.258999986
#> 55   9.533960 0.3777005726 0.6860220742 0.38480090 0.11102878 0.201999931
#> 56   8.106522 0.9152705607 0.1164152857 0.11095753 0.08064474 0.212551820
#> 57  10.366087 0.8287820488 0.1184740260 0.32069635 0.31668340 0.694759363
#> 58  16.214461 0.9444349702 0.9365332325 0.43044472 0.94069411 0.633837986
#> 59  12.238676 0.1867862260 0.9565980940 0.54563143 0.57530196 0.237179739
#> 60  17.689560 0.4802156268 0.9164127647 0.51170095 0.71697276 0.138801176
#> 61   2.320966 0.2689777846 0.0007479317 0.41833439 0.15742279 0.126832403
#> 62  13.735266 0.4354642837 0.5360080854 0.36761624 0.31375840 0.706044929
#> 63  15.009681 0.0625840209 0.5439596830 0.86482505 0.83520194 0.608523829
#> 64  20.744012 0.3366812817 0.8265306321 0.87867310 0.78557594 0.477584738
#> 65  12.108408 0.5619351428 0.3660433823 0.39835940 0.45182194 0.250936965
#> 66  14.271105 0.6923129486 0.1326507980 0.84774744 0.78260741 0.218694113
#> 67  11.190654 0.1265436679 0.7145354096 0.36338261 0.53980688 0.508063156
#> 68  16.025768 0.5177793652 0.9832966432 0.68576940 0.43944972 0.190326528
#> 69  24.789005 0.6388091939 0.3776578177 0.98700774 0.84541431 0.951850409
#> 70   9.185519 0.0463474730 0.2209118321 0.50158703 0.46888338 0.823391459
#> 71   7.108708 0.2081563191 0.1640034495 0.46008040 0.28993239 0.618096247
#> 72  12.394803 0.2881255990 0.1692087383 0.37472480 0.70474861 0.648348246
#> 73  15.439979 0.4617162470 0.7853772915 0.90474155 0.18388422 0.262976428
#> 74  13.130026 0.4069326075 0.6049174396 0.98160203 0.09183642 0.097967924
#> 75  19.067879 0.7198746817 0.9837067495 0.24622017 0.58417129 0.782056808
#> 76  18.571389 0.4783525530 0.6985027203 0.54063990 0.65038579 0.681180339
#> 77  18.058464 0.8916681909 0.3719108894 0.47872891 0.60925668 0.674633251
#> 78   4.790314 0.0246532848 0.1542579043 0.42178781 0.44500275 0.037436911
#> 79  23.331516 0.6351153189 0.5481393081 0.04448537 0.64214677 0.790769406
#> 80  17.270832 0.3908880984 0.7231449513 0.63429824 0.72586831 0.392556404
#> 81  16.783845 0.5376661033 0.5836053693 0.36974378 0.72212535 0.162541629
#> 82  16.809608 0.7452690604 0.5157998072 0.59968251 0.25374419 0.936242811
#> 83   8.672745 0.1479747796 0.7217902243 0.77612701 0.36192930 0.048533057
#> 84  15.491656 0.1909503259 0.6034291473 0.73234764 0.97781434 0.179092183
#> 85   8.849248 0.3254722552 0.1843186312 0.19070342 0.42697828 0.165738771
#> 86   9.423387 0.0394796270 0.7239926779 0.13016036 0.25472619 0.647701826
#> 87  14.954767 0.5934679159 0.8440855062 0.81376928 0.06636500 0.459419391
#> 88  14.940362 0.6241817074 0.8355207876 0.68256784 0.12848377 0.579249768
#> 89  22.904172 0.5443477246 0.7064004592 0.23747160 0.99472952 0.446649317
#> 90   4.553068 0.0027585044 0.7108768146 0.28378503 0.21110891 0.299774076
#> 91  13.588892 0.9419699602 0.8855582720 0.64001952 0.35364491 0.938302146
#> 92   7.685669 0.1348288120 0.0441270615 0.16040167 0.21688171 0.571468831
#> 93  19.438779 0.6910544815 0.8454461338 0.81100623 0.54850378 0.458284137
#> 94  16.092476 0.7715629931 0.1065962985 0.99087681 0.66833139 0.399829024
#> 95  16.937315 0.6247570168 0.3522469536 0.58727770 0.87004411 0.353819617
#> 96  17.355054 0.8089950338 0.2592346051 0.58606783 0.76391437 0.714789014
#> 97   9.297988 0.6508365592 0.2403352042 0.49607992 0.12195164 0.641367927
#> 98  11.444188 0.2195732470 0.9544442415 0.63417187 0.19763729 0.621912588
#> 99  15.596308 0.7400014549 0.2873893159 0.22290774 0.31125073 0.934812914
#> 100 22.280968 0.4325508266 0.8760279566 0.46124819 0.88522120 0.853565969
#>              x6          x7         x8           x9        x10
#> 1   0.034242923 0.708089868 0.35914147 0.0610777270 0.22758428
#> 2   0.596107564 0.486425376 0.49329318 0.8551026261 0.43574008
#> 3   0.285908719 0.213685652 0.81126283 0.5519285847 0.83989131
#> 4   0.293156812 0.882510883 0.22718654 0.4651124007 0.64320816
#> 5   0.012705019 0.948371802 0.63928203 0.3378157120 0.32491792
#> 6   0.422165087 0.246071426 0.03615783 0.2795106776 0.71561663
#> 7   0.120269010 0.572805416 0.80556467 0.7345445950 0.73703060
#> 8   0.504527518 0.369914019 0.29122099 0.9650976271 0.26642593
#> 9   0.007521316 0.600617972 0.93391782 0.1898512354 0.69748546
#> 10  0.433976179 0.753366186 0.95327126 0.5203839843 0.13945654
#> 11  0.342134999 0.590422563 0.26271290 0.0611708523 0.37207212
#> 12  0.776205650 0.188571413 0.09173083 0.6422129232 0.28489947
#> 13  0.737612485 0.106413301 0.69908908 0.8835501899 0.02540059
#> 14  0.626222877 0.047179515 0.51425558 0.3463421324 0.15171488
#> 15  0.955695485 0.186753943 0.13663220 0.6762772684 0.59742975
#> 16  0.172546203 0.376719295 0.38715997 0.0125030046 0.54510263
#> 17  0.545021115 0.484789640 0.36758738 0.4416310841 0.22442658
#> 18  0.538295767 0.892354126 0.78598531 0.0343003345 0.26576790
#> 19  0.787723745 0.208301702 0.63139955 0.4674016791 0.25383709
#> 20  0.396561470 0.085427377 0.14193047 0.0041771673 0.96478442
#> 21  0.630536117 0.070383406 0.35941601 0.9448859505 0.46534421
#> 22  0.972894424 0.633558517 0.04530535 0.6375728617 0.73293406
#> 23  0.039641326 0.655676758 0.59222664 0.5741242785 0.05137181
#> 24  0.517838627 0.033208641 0.69524889 0.6965280552 0.90342846
#> 25  0.659225761 0.711370954 0.22926428 0.4074910728 0.02295865
#> 26  0.132956168 0.321289820 0.30945442 0.6835767005 0.31565006
#> 27  0.618167999 0.829285606 0.24049369 0.7298645473 0.54870292
#> 28  0.557701408 0.632464110 0.26868906 0.9848905541 0.96519751
#> 29  0.050808229 0.704465387 0.17299500 0.2793988590 0.90850278
#> 30  0.117315194 0.766360630 0.25847356 0.0828559217 0.11078018
#> 31  0.924157008 0.947811504 0.11520784 0.4628402672 0.51810978
#> 32  0.691690837 0.696576672 0.28472149 0.6932067138 0.28909726
#> 33  0.585145741 0.619186329 0.83799210 0.2473595894 0.74913218
#> 34  0.377643411 0.862922556 0.88198908 0.6925247158 0.40997901
#> 35  0.047371062 0.266061272 0.18454526 0.4830463436 0.18086288
#> 36  0.517291918 0.243575299 0.73558081 0.3417504118 0.68246246
#> 37  0.352946862 0.693149599 0.14363070 0.1045486750 0.41737379
#> 38  0.579649796 0.596710179 0.14520207 0.8336928731 0.86301916
#> 39  0.416965917 0.499558397 0.34470076 0.6475775053 0.04587891
#> 40  0.493167195 0.583622808 0.08443853 0.5942029294 0.15718454
#> 41  0.170384483 0.874281158 0.70761097 0.3040392618 0.14285383
#> 42  0.821086831 0.946794922 0.71313778 0.6325784195 0.55167329
#> 43  0.166999853 0.243589720 0.75084571 0.2689357174 0.25884974
#> 44  0.252969695 0.853858619 0.13020640 0.8448705720 0.87985808
#> 45  0.109633006 0.900220604 0.67633549 0.0168415420 0.75819330
#> 46  0.752872009 0.094513787 0.24641105 0.9307651916 0.36220957
#> 47  0.414518373 0.303107897 0.79306630 0.1548446179 0.20261637
#> 48  0.164274446 0.929899182 0.54470920 0.6314759830 0.25484556
#> 49  0.443809730 0.115665130 0.43970373 0.0353378267 0.87141381
#> 50  0.511488160 0.699254848 0.71063872 0.3781803669 0.83220172
#> 51  0.670928738 0.985215642 0.44290613 0.7118906244 0.40738102
#> 52  0.426001019 0.450633260 0.21228880 0.6114282946 0.23299247
#> 53  0.404537401 0.350622531 0.32384877 0.6979654906 0.24376099
#> 54  0.836621791 0.424080716 0.44632980 0.7862614230 0.70039763
#> 55  0.047172359 0.639840494 0.93305562 0.5415839397 0.30996131
#> 56  0.742328340 0.900150308 0.90837021 0.9925867668 0.26425059
#> 57  0.628580088 0.200577693 0.66036661 0.8578269284 0.31954179
#> 58  0.647300746 0.290988813 0.88610383 0.6013091863 0.99704871
#> 59  0.611986624 0.764279517 0.16331195 0.4026571792 0.65303878
#> 60  0.432625725 0.317979577 0.57190554 0.8151451547 0.06493345
#> 61  0.262366948 0.003992743 0.61654453 0.8046097308 0.73452569
#> 62  0.632444909 0.282399718 0.03126944 0.4877454117 0.92764269
#> 63  0.168103796 0.374801098 0.59942496 0.9299545551 0.57508326
#> 64  0.538680434 0.105743600 0.33625236 0.5762514903 0.37320879
#> 65  0.406551675 0.448441658 0.73176304 0.6472766520 0.06112425
#> 66  0.434686203 0.778510033 0.07953886 0.0003732827 0.49847027
#> 67  0.171017120 0.054072363 0.48412030 0.1010017192 0.23735809
#> 68  0.960915104 0.996568874 0.74783007 0.3287396773 0.83769274
#> 69  0.114367873 0.676393326 0.87083452 0.0801869438 0.97321203
#> 70  0.845674331 0.998511230 0.28930782 0.8951015933 0.75409960
#> 71  0.236174762 0.897463962 0.91057811 0.3170280487 0.66825403
#> 72  0.103356488 0.882108086 0.92159107 0.8556554778 0.78922504
#> 73  0.950230394 0.163734372 0.08825122 0.1731632191 0.90219313
#> 74  0.922517024 0.708412708 0.12089556 0.1927078371 0.15004643
#> 75  0.475061484 0.674637960 0.99018488 0.4840292903 0.20470819
#> 76  0.555780485 0.547723550 0.50100064 0.7315694520 0.49782261
#> 77  0.923358585 0.090754932 0.61106268 0.8990699153 0.10081324
#> 78  0.971798369 0.162182079 0.66534366 0.9558014688 0.91095628
#> 79  0.401708246 0.426679686 0.18814570 0.8032109921 0.08995544
#> 80  0.196741019 0.986763639 0.82722021 0.1786617143 0.01899216
#> 81  0.462164629 0.353265165 0.26879851 0.9008497326 0.05938888
#> 82  0.038918576 0.139789863 0.16585021 0.7366295452 0.81893303
#> 83  0.096510025 0.767528786 0.68874944 0.8194924979 0.34656378
#> 84  0.212969644 0.399658867 0.38180739 0.0589272704 0.72287816
#> 85  0.663871369 0.961870601 0.09616278 0.4503492787 0.78193040
#> 86  0.956607925 0.490030340 0.84577716 0.2624837207 0.54147464
#> 87  0.355419232 0.110987839 0.70473588 0.7271480360 0.35262759
#> 88  0.885159231 0.446340722 0.10877450 0.6382283943 0.45124294
#> 89  0.697324503 0.054131986 0.08702338 0.0186577390 0.84914334
#> 90  0.792573904 0.137287467 0.41453144 0.6244785439 0.02352309
#> 91  0.724002579 0.913146948 0.53844482 0.3332280456 0.70290222
#> 92  0.488631977 0.896673730 0.31123061 0.1474314202 0.10678558
#> 93  0.015908889 0.837059021 0.53906593 0.3760554059 0.11756390
#> 94  0.785463817 0.141332611 0.73425765 0.3223124736 0.78057825
#> 95  0.206332984 0.361111816 0.95616994 0.7208354436 0.95605403
#> 96  0.120416545 0.355863324 0.66917523 0.6672515746 0.60843793
#> 97  0.704068359 0.377924406 0.85128665 0.2099642872 0.34120736
#> 98  0.962544315 0.795774869 0.12637500 0.1922181458 0.20560801
#> 99  0.396618249 0.696234760 0.82306710 0.4727696877 0.50951772
#> 100 0.954556089 0.140891456 0.78722637 0.5780899401 0.80233188