Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. These data were originally described in Friedman (1991) and Breiman (1996). For details, see sklearn.datasets.make_friedman1.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

References

Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The Annals of Statistics 19 (1), pages 1-67.

Examples

gen_friedman()
#>             y          x1          x2          x3          x4          x5
#> 1   16.976501 0.631675954 0.871071756 0.452977981 0.413325371 0.567986187
#> 2   16.500167 0.858684332 0.412470120 0.390449678 0.309109204 0.829878627
#> 3   15.601748 0.217330578 0.421157449 0.033231578 0.433165770 0.837966824
#> 4   17.533231 0.680877398 0.414154296 0.134868340 0.564376870 0.316282573
#> 5    9.293548 0.983244197 0.093391256 0.196743613 0.141799615 0.635765992
#> 6    8.301749 0.030674560 0.497356236 0.947646512 0.014908865 0.737200767
#> 7   16.935036 0.804780768 0.731981588 0.718363146 0.161310408 0.964792164
#> 8   22.427294 0.715037743 0.613705293 0.893818433 0.933948228 0.011436522
#> 9   18.144511 0.907574862 0.223045166 0.212348065 0.989781432 0.134624908
#> 10  21.059047 0.938068524 0.686288204 0.814325405 0.527960867 0.968026098
#> 11  15.889141 0.151185663 0.831157442 0.902057439 0.583458502 0.609258221
#> 12  16.839157 0.152444293 0.691116907 0.256232996 0.864358238 0.766976811
#> 13  13.314227 0.255576581 0.975898861 0.137206498 0.101652046 0.526893531
#> 14  16.237616 0.283447837 0.704952119 0.277733215 0.481040020 0.876323058
#> 15  17.663455 0.748991321 0.947099699 0.492040504 0.964333974 0.014011946
#> 16  13.764638 0.766947272 0.181066280 0.078335340 0.501717729 0.230233608
#> 17  17.146282 0.670559488 0.736486087 0.144985345 0.165321273 0.597705806
#> 18  10.346533 0.879458235 0.945342788 0.472540217 0.298373228 0.458441418
#> 19  11.136658 0.340776464 0.904966528 0.257049969 0.011713040 0.295522292
#> 20  11.455153 0.728911526 0.381506856 0.179182523 0.180261683 0.016562122
#> 21   7.747713 0.199192752 0.182997522 0.382926679 0.289993959 0.656604981
#> 22  20.128720 0.673630706 0.905124306 0.850867569 0.534144673 0.557108087
#> 23  18.208937 0.298157707 0.530918019 0.667573967 0.977064086 0.608102854
#> 24  17.187613 0.512791863 0.350287767 0.082814153 0.739822852 0.198038888
#> 25  15.624657 0.950181397 0.178702033 0.674235711 0.544856545 0.881971820
#> 26  17.589743 0.982253800 0.275918534 0.846229583 0.578956767 0.390852148
#> 27  20.378074 0.568654864 0.569132434 0.350533324 0.981898185 0.334659056
#> 28  10.643517 0.333497354 0.675749324 0.361356799 0.212695696 0.319730877
#> 29  15.485922 0.322848451 0.331371719 0.602118141 0.788985869 0.836390581
#> 30   8.277770 0.007553466 0.750428062 0.040020139 0.178497395 0.411190092
#> 31  20.113125 0.384506829 0.592706370 0.834609390 0.825227508 0.604128249
#> 32  13.808075 0.338630203 0.856547253 0.418496298 0.451200804 0.274619463
#> 33   9.538215 0.314667469 0.061906121 0.195764970 0.668047984 0.064784759
#> 34   7.250367 0.656836455 0.115496995 0.390102384 0.297654604 0.347705300
#> 35  24.310882 0.757793591 0.352848354 0.156595781 0.993813467 0.929786792
#> 36  13.682749 0.327556049 0.875483254 0.414102377 0.165837998 0.819050186
#> 37  19.780169 0.530115247 0.875364512 0.762798940 0.705115503 0.286503908
#> 38  13.119034 0.074512150 0.941768235 0.876485183 0.673669763 0.261336010
#> 39  18.287978 0.817752601 0.376784203 0.036657023 0.222991786 0.718754775
#> 40  16.625020 0.518982120 0.847237778 0.334363921 0.488637134 0.257681591
#> 41  13.262126 0.233352149 0.509520798 0.905732569 0.527135487 0.210310126
#> 42  10.918240 0.834256413 0.112036514 0.133976916 0.350815791 0.350432169
#> 43  14.987739 0.380075325 0.729417264 0.781516820 0.132279771 0.902333182
#> 44  15.504113 0.930803411 0.467032204 0.785141123 0.256642200 0.309700339
#> 45   8.853122 0.849447894 0.120294221 0.859646334 0.154152506 0.305555001
#> 46   8.742356 0.414229673 0.342575222 0.729831379 0.115233117 0.423189257
#> 47   8.929244 0.057264397 0.599118719 0.205685571 0.140003619 0.929742931
#> 48   8.154509 0.059754730 0.784022792 0.582958510 0.202737621 0.914917475
#> 49  11.454988 0.811985419 0.257417285 0.856735481 0.008718245 0.561127441
#> 50   7.721815 0.260581260 0.827362668 0.418924719 0.089823189 0.088866886
#> 51  15.143780 0.378367144 0.688762853 0.131348478 0.432240363 0.212315380
#> 52   8.595650 0.013552682 0.926721674 0.581191823 0.336882074 0.935887425
#> 53  17.246627 0.783056194 0.421886894 0.345190956 0.503531415 0.613621308
#> 54  13.979214 0.742456208 0.687150940 0.588961160 0.339627716 0.074835765
#> 55  15.473307 0.979803348 0.141029086 0.664016492 0.762551563 0.656373372
#> 56  16.926664 0.866342029 0.916405894 0.675540888 0.801727324 0.457001157
#> 57  10.165943 0.409667948 0.215428264 0.753078074 0.387576644 0.484638488
#> 58   9.037749 0.558474664 0.083192638 0.464889758 0.350203032 0.817837341
#> 59  19.175045 0.422208135 0.875313197 0.058783069 0.567895770 0.054295819
#> 60  15.997342 0.076287633 0.930295785 0.170241066 0.869748494 0.586870526
#> 61  13.654272 0.289538463 0.586552708 0.738607484 0.321028014 0.816269561
#> 62  15.820964 0.542793660 0.709899094 0.389952763 0.375379004 0.502505699
#> 63  19.220668 0.372325180 0.582539384 0.923390516 0.759172606 0.369191800
#> 64  14.837755 0.154248002 0.864648741 0.619746975 0.651784434 0.782610946
#> 65   8.479238 0.225768549 0.628813348 0.249175259 0.126718372 0.328422425
#> 66  21.571506 0.651690922 0.546482912 0.494515233 0.990661174 0.523642704
#> 67   6.998764 0.022850175 0.584626310 0.873836140 0.179837878 0.366222899
#> 68  12.559482 0.642870787 0.208829887 0.903041041 0.072073469 0.852402804
#> 69  18.251242 0.296406037 0.563380719 0.368478296 0.932403311 0.699552755
#> 70  17.530375 0.867287372 0.216257367 0.178094242 0.579822404 0.848219468
#> 71   9.668344 0.676768060 0.461091608 0.343458120 0.079248077 0.020703070
#> 72  15.475823 0.632379413 0.694762170 0.661137730 0.413864980 0.202702707
#> 73   7.629884 0.330350673 0.106541197 0.916503269 0.289945686 0.011651113
#> 74  12.827223 0.318142316 0.667146289 0.696450128 0.529771638 0.143643437
#> 75  15.957321 0.548335272 0.050607361 0.994953345 0.556043397 0.939561758
#> 76  21.122006 0.541488892 0.953741503 0.138677525 0.695784054 0.320812795
#> 77   8.797551 0.721150326 0.006840569 0.924137221 0.319116057 0.362254124
#> 78  14.150114 0.810855828 0.417220214 0.835206240 0.210277536 0.221620744
#> 79  23.862369 0.703562062 0.796775954 0.192128665 0.738352766 0.960866016
#> 80  10.739785 0.764227478 0.315939393 0.460157309 0.265635962 0.246862028
#> 81  15.575486 0.125109596 0.770906801 0.951810396 0.491825133 0.711263651
#> 82  21.742240 0.689516194 0.829238424 0.706124878 0.664201712 0.912496795
#> 83  20.385541 0.393511047 0.666325775 0.633540899 0.820764035 0.874064176
#> 84  12.175272 0.215407399 0.420132505 0.813795059 0.652919258 0.163809486
#> 85  16.142396 0.137166417 0.588639467 0.652264694 0.874402949 0.889067861
#> 86  12.396822 0.870763201 0.773689958 0.646244966 0.093679758 0.503140346
#> 87  15.543601 0.621258410 0.829953831 0.319102112 0.482637174 0.038597407
#> 88   6.800806 0.162131972 0.597335299 0.270391211 0.014733215 0.527000012
#> 89  13.728389 0.287192781 0.508753046 0.923526762 0.491288682 0.156567883
#> 90  16.630244 0.751478297 0.953530994 0.693353237 0.649227770 0.301724473
#> 91  15.070963 0.830836474 0.282007494 0.083005948 0.307249689 0.361677159
#> 92  15.603323 0.973174233 0.622810306 0.527416503 0.180558371 0.865405990
#> 93  10.809361 0.358959578 0.390367604 0.266123049 0.393414928 0.282301308
#> 94  18.259126 0.216115972 0.539020203 0.001644427 0.976964086 0.005082714
#> 95  15.724433 0.360427620 0.409314780 0.294683304 0.610500163 0.856790501
#> 96  16.354863 0.404407139 0.798976548 0.862926559 0.381593765 0.271634173
#> 97   6.896775 0.079218751 0.918491453 0.940549014 0.014105887 0.145797194
#> 98  12.131737 0.339745495 0.233693007 0.428051462 0.903655064 0.102115871
#> 99   7.563889 0.750417240 0.063670425 0.768474457 0.205951801 0.533231065
#> 100 13.202971 0.686906547 0.454122943 0.400372445 0.281985777 0.364768770
#>              x6          x7          x8         x9        x10
#> 1   0.161107766 0.207619793 0.321073955 0.93953274 0.11684485
#> 2   0.076696777 0.221408810 0.362764165 0.79414422 0.34944642
#> 3   0.578651103 0.606859197 0.176077551 0.79948664 0.01188524
#> 4   0.419518960 0.584720683 0.916456710 0.08746218 0.12155954
#> 5   0.613769640 0.918332098 0.894107062 0.56514734 0.75178880
#> 6   0.265238645 0.881183226 0.573668035 0.70270337 0.96560767
#> 7   0.440914714 0.556675691 0.457651958 0.87297605 0.39224583
#> 8   0.227210416 0.462672091 0.024690456 0.66794293 0.68724588
#> 9   0.558987414 0.721355593 0.715030433 0.70011243 0.09316872
#> 10  0.272494360 0.257448323 0.401939621 0.38127742 0.83136247
#> 11  0.327549207 0.324307371 0.455893323 0.31024099 0.19730209
#> 12  0.854709671 0.912844302 0.503065157 0.28701246 0.91480437
#> 13  0.319968575 0.245129285 0.085890463 0.38376872 0.57612077
#> 14  0.539391186 0.712612598 0.866045458 0.90616914 0.66910398
#> 15  0.545354954 0.420213789 0.838262249 0.14103002 0.21374972
#> 16  0.372133990 0.432053106 0.664676750 0.58666016 0.40908570
#> 17  0.572657129 0.586230555 0.669527009 0.01787209 0.69883866
#> 18  0.119764886 0.842323753 0.099777168 0.86167619 0.83038949
#> 19  0.488243777 0.816039367 0.609593143 0.11124351 0.73567816
#> 20  0.349650876 0.819434965 0.254921244 0.76276629 0.42189910
#> 21  0.795744601 0.018551452 0.730956452 0.60364026 0.23025507
#> 22  0.876236638 0.049971209 0.155718811 0.08012965 0.36423207
#> 23  0.166279150 0.735469294 0.738155313 0.50147699 0.35660879
#> 24  0.858372532 0.584397572 0.489948607 0.91804126 0.45992293
#> 25  0.847948894 0.325493190 0.637033261 0.42646382 0.43455210
#> 26  0.668455952 0.312400682 0.250640373 0.38358221 0.92680333
#> 27  0.700480191 0.944765502 0.508218075 0.59167270 0.15433466
#> 28  0.335669094 0.825937907 0.517497334 0.77160028 0.19964521
#> 29  0.038844270 0.348708753 0.682955612 0.02217775 0.94651502
#> 30  0.245959383 0.062952528 0.403085410 0.85178182 0.86897015
#> 31  0.761252699 0.353579039 0.393752639 0.90701903 0.26450129
#> 32  0.353009412 0.074011315 0.437797458 0.87312466 0.13938073
#> 33  0.001460670 0.092309304 0.531889361 0.46911221 0.29723975
#> 34  0.775693191 0.006452727 0.336770533 0.79542085 0.63525833
#> 35  0.378261218 0.059072361 0.441153008 0.20111579 0.25159051
#> 36  0.759531138 0.297545365 0.884616688 0.94581096 0.19642292
#> 37  0.685180195 0.131467561 0.824858066 0.09443136 0.31385977
#> 38  0.681939323 0.789758257 0.932677008 0.24645788 0.54341242
#> 39  0.062953055 0.927160287 0.259854890 0.95402948 0.55335344
#> 40  0.373360688 0.029847582 0.800718169 0.89330293 0.53256034
#> 41  0.996519804 0.511316691 0.942573381 0.41904545 0.50923538
#> 42  0.083126765 0.574793985 0.538756164 0.28729313 0.84850562
#> 43  0.735013397 0.848474107 0.365843052 0.36884323 0.37555758
#> 44  0.730195088 0.821220627 0.363562454 0.38786375 0.96209655
#> 45  0.020147586 0.155607012 0.433470801 0.18218565 0.80706843
#> 46  0.349610718 0.572511432 0.108151437 0.33624012 0.30081576
#> 47  0.152629047 0.818633312 0.437166714 0.75967797 0.33444318
#> 48  0.605226003 0.679017249 0.515897860 0.28439333 0.53642806
#> 49  0.539751902 0.164685660 0.171859787 0.89255698 0.31363215
#> 50  0.397091195 0.670569756 0.585377050 0.19779494 0.71154964
#> 51  0.614781228 0.248419169 0.352279752 0.13256540 0.47476122
#> 52  0.857612227 0.675538447 0.945126421 0.86094123 0.77470260
#> 53  0.471249504 0.235065642 0.188045381 0.84508093 0.32580347
#> 54  0.947907970 0.645181842 0.575135570 0.40028734 0.18098013
#> 55  0.679192248 0.669623616 0.512769716 0.90264532 0.32794292
#> 56  0.637725082 0.016759021 0.842446565 0.73950951 0.81023964
#> 57  0.600657429 0.796105374 0.582868275 0.37016346 0.33796791
#> 58  0.713750661 0.900793898 0.608004504 0.64909395 0.32961305
#> 59  0.713987009 0.772946005 0.898918698 0.07608635 0.30523902
#> 60  0.122523598 0.826322920 0.458950059 0.44117554 0.76829471
#> 61  0.398549309 0.309410169 0.065268878 0.35359261 0.36794106
#> 62  0.145626288 0.608956100 0.627678412 0.86841208 0.32949566
#> 63  0.649903436 0.687768716 0.675620860 0.91761154 0.94183182
#> 64  0.218477824 0.773477252 0.809933146 0.40923497 0.58506563
#> 65  0.719168668 0.105931549 0.766353937 0.73852222 0.06533017
#> 66  0.024382079 0.421726623 0.717581614 0.33234915 0.26227351
#> 67  0.052096129 0.082307472 0.992746690 0.73052220 0.02540879
#> 68  0.422759339 0.951485484 0.661980039 0.58804618 0.96098424
#> 69  0.154971749 0.950662453 0.805575543 0.60094614 0.98733969
#> 70  0.485582996 0.916782238 0.938234730 0.36655052 0.29930503
#> 71  0.562938660 0.598956770 0.326024584 0.70583623 0.56763074
#> 72  0.500579153 0.478949799 0.659606672 0.81160390 0.83253549
#> 73  0.699928174 0.255624072 0.706133065 0.23268532 0.60652998
#> 74  0.176226656 0.665307347 0.577219328 0.44985302 0.27784112
#> 75  0.901931266 0.868160985 0.546723743 0.32682807 0.91055433
#> 76  0.473941420 0.019240447 0.002589624 0.16352290 0.52874481
#> 77  0.312441811 0.457007555 0.898472774 0.57034617 0.24437921
#> 78  0.616185558 0.087150355 0.624914761 0.89071674 0.38374880
#> 79  0.009891189 0.401458647 0.231384394 0.90559741 0.47999151
#> 80  0.132968071 0.668871341 0.380589359 0.99620026 0.19071918
#> 81  0.309769478 0.465531803 0.863866784 0.12184399 0.07896498
#> 82  0.851714578 0.297438596 0.256893829 0.56172501 0.40135647
#> 83  0.954040427 0.864228091 0.172863220 0.45632057 0.59932214
#> 84  0.341569389 0.822225519 0.672400513 0.01668140 0.48029403
#> 85  0.335157020 0.068215986 0.850826806 0.70217682 0.68324932
#> 86  0.120605124 0.740097270 0.165290598 0.85731817 0.61093398
#> 87  0.869803404 0.274550612 0.165004834 0.21150869 0.62391872
#> 88  0.548132747 0.027441496 0.893007161 0.28038652 0.73824212
#> 89  0.920355980 0.855052414 0.556374932 0.54995071 0.44804424
#> 90  0.824874862 0.922223949 0.692704028 0.83684463 0.76759089
#> 91  0.888955155 0.262072365 0.964542825 0.11575496 0.62441113
#> 92  0.150484454 0.840030419 0.307424065 0.66900638 0.39469965
#> 93  0.495364265 0.398318323 0.924772107 0.65580030 0.97717260
#> 94  0.350593768 0.790977478 0.956345581 0.90622495 0.01614503
#> 95  0.042435116 0.139606323 0.100889059 0.96109988 0.27091998
#> 96  0.741039617 0.780601638 0.614621850 0.67526345 0.53697986
#> 97  0.357732598 0.202104322 0.796255789 0.22763737 0.87989101
#> 98  0.757641173 0.865707470 0.200830390 0.92245054 0.17206503
#> 99  0.110809144 0.838882552 0.565201156 0.97762240 0.61076208
#> 100 0.552645630 0.390805527 0.704040335 0.21619488 0.70113720