Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. See mlbench.friedman1 for details and references.

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples
Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features
Integer specifying the number of features to generate. Default is 10.

n_bins
Integer specying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Settinig to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma
Numeric specifying the standard deviation of the noise.

seed
Integer specying the random seed. If NULL (the default) the results will be different each time the function is run.

Note

This function is mostly used for internal testing.

Examples

gen_friedman()

#>             y          x1          x2         x3           x4          x5
#> 1   10.742781 0.080750138 0.805680031 0.34929905 0.6660920116 0.333823317
#> 2   18.310349 0.834333037 0.814051310 0.94731827 0.2039006804 0.749272652
#> 3   18.598813 0.600760886 0.403911002 0.21609998 0.5485914391 0.913158011
#> 4   14.771906 0.157208442 0.218431009 0.03209271 0.8275014143 0.204329063
#> 5    6.127266 0.007399441 0.418361402 0.14531584 0.1275576509 0.471755611
#> 6   15.350760 0.466393497 0.668870748 0.85438389 0.2611861178 0.385183100
#> 7   13.693016 0.497777389 0.507650282 0.21314931 0.2884170276 0.392751929
#> 8    8.644180 0.289767245 0.660359306 0.21031074 0.0147690158 0.218338430
#> 9   22.153039 0.732881987 0.511791314 0.03952069 0.8537237272 0.007208941
#> 10  19.003557 0.772521511 0.835552437 0.94477480 0.4140478286 0.363712143
#> 11  20.616931 0.874600661 0.708781161 0.24492799 0.7510138457 0.526106730
#> 12  17.878887 0.174940627 0.874205941 0.78112257 0.7190139582 0.844163393
#> 13   6.727036 0.034241333 0.011479538 0.28823717 0.5153553013 0.135991171
#> 14  22.202318 0.320385731 0.888249568 0.87535791 0.9748362575 0.376467331
#> 15  15.257706 0.402328238 0.996346919 0.29575009 0.3607775616 0.263308648
#> 16  11.806208 0.195669835 0.500191501 0.98352541 0.1703955070 0.497852970
#> 17  12.576229 0.403538117 0.358967024 0.58983756 0.7831406700 0.019525251
#> 18   3.640001 0.063661457 0.774913022 0.75915838 0.0309238962 0.087183810
#> 19  21.553621 0.388701313 0.584475251 0.83607531 0.7894203586 0.995962664
#> 20  17.176051 0.975547835 0.633976371 0.76281947 0.1666036777 0.942132205
#> 21   8.019852 0.289892295 0.858666155 0.41726993 0.0287125653 0.124373451
#> 22  20.385897 0.678380427 0.566894339 0.13807484 0.7859511641 0.115189302
#> 23  19.170565 0.735319599 0.252997019 0.08084496 0.8253078635 0.366688104
#> 24  17.908716 0.195956733 0.918803215 0.65598263 0.9652404180 0.471040036
#> 25  12.497046 0.980539675 0.867350205 0.60200386 0.3786411788 0.784172446
#> 26  10.094577 0.741521529 0.248538697 0.65699583 0.1743863276 0.490392813
#> 27   7.786343 0.051446276 0.402881212 0.32931716 0.6037504023 0.101950955
#> 28  26.674868 0.530212464 0.769630176 0.97947422 0.8054637336 0.943459724
#> 29   7.314714 0.695823879 0.119485375 0.71518613 0.0366794092 0.706139502
#> 30  16.627422 0.688556003 0.194694962 0.87263030 0.7329106666 0.502264237
#> 31   9.568710 0.031230325 0.164569248 0.98328375 0.2152726583 0.529934776
#> 32   7.627482 0.225562535 0.663206581 0.21856299 0.0160031442 0.258426348
#> 33  12.575161 0.300830806 0.856575005 0.66453006 0.1286025783 0.711753417
#> 34  21.579333 0.636465615 0.926546448 0.38956404 0.6860631041 0.968462550
#> 35  21.280613 0.479024550 0.552377595 0.04606364 0.6419412645 0.661077790
#> 36  14.657949 0.432171258 0.577065694 0.61691456 0.3273385202 0.796250590
#> 37  14.486016 0.706433838 0.687447746 0.59847499 0.3874819812 0.073209132
#> 38  18.273714 0.948576576 0.244718230 0.40685363 0.7075003954 0.884742884
#> 39  13.952862 0.180338768 0.044617158 0.85832815 0.6976554273 0.826700377
#> 40  18.814662 0.216899876 0.909854557 0.51768118 0.9256567662 0.725763959
#> 41  12.707189 0.680162918 0.070681219 0.97929341 0.4590177245 0.367329022
#> 42  23.233417 0.498845611 0.996891473 0.01701569 0.5959542021 0.523428417
#> 43  14.174256 0.641679348 0.611852417 0.67344783 0.1650109896 0.469878913
#> 44   9.934131 0.660284349 0.172558846 0.37126988 0.5669097845 0.098566386
#> 45  17.743156 0.096024158 0.909440965 0.91801064 0.8980773711 0.516034910
#> 46   9.658104 0.765600164 0.037451167 0.67797809 0.5944722812 0.449733827
#> 47  21.473474 0.769674804 0.593553790 0.66515246 0.8316899191 0.549461548
#> 48  17.278336 0.990712312 0.236977555 0.75604109 0.5934083650 0.649506190
#> 49  12.377497 0.970520903 0.906297267 0.54283715 0.7789706336 0.159034826
#> 50  15.157518 0.389182761 0.818872984 0.23928810 0.3977716470 0.293750792
#> 51  21.367756 0.461186465 0.699829357 0.50889357 0.8498827741 0.891284867
#> 52  12.562040 0.315241752 0.220000329 0.41726437 0.7418455814 0.588078626
#> 53  11.442961 0.174675894 0.727990938 0.72694885 0.3177902193 0.633231563
#> 54   6.374787 0.531573541 0.217084462 0.63768555 0.1116802238 0.259509476
#> 55   9.342092 0.493637016 0.456230198 0.39640996 0.1010953654 0.318212727
#> 56  21.114910 0.779308626 0.332799758 0.95948261 0.8000977482 0.339253881
#> 57  11.990060 0.204178343 0.568352669 0.29865803 0.3799467422 0.760761139
#> 58   9.950614 0.713397279 0.252205725 0.05020117 0.0527009717 0.010872464
#> 59  11.410300 0.065216112 0.464013567 0.57618742 0.9865086977 0.063137339
#> 60  17.415185 0.354206799 0.917660507 0.21790581 0.6041739669 0.263841792
#> 61  13.122087 0.825199421 0.972844218 0.12585627 0.1487827229 0.593836589
#> 62  17.177174 0.273818245 0.819082447 0.93815269 0.5387601722 0.297560029
#> 63  13.364586 0.570044951 0.902923798 0.80127513 0.1257892270 0.094622772
#> 64  20.667598 0.335719081 0.581366044 0.75805362 0.9631817536 0.761255360
#> 65  14.013136 0.596262789 0.773008481 0.53256516 0.0469350144 0.772476848
#> 66   9.200611 0.191518032 0.995123026 0.54680477 0.1614091825 0.401990393
#> 67  21.400552 0.947763938 0.710971250 0.09592650 0.9364874261 0.059082953
#> 68  17.219100 0.542480409 0.214942596 0.38834975 0.9706813623 0.750247884
#> 69  16.732216 0.544603393 0.291757630 0.17235189 0.7102740526 0.528322355
#> 70  16.831102 0.278597154 0.721759729 0.69072585 0.8850909304 0.275805118
#> 71  20.982707 0.446702469 0.866615703 0.67520850 0.9768753166 0.261237550
#> 72   9.663527 0.371511184 0.238453106 0.94629485 0.0348996436 0.543831169
#> 73   7.150435 0.028060974 0.004496308 0.19621952 0.4387572692 0.199000779
#> 74  23.819858 0.465987191 0.943516464 0.96863750 0.6444697606 0.645054772
#> 75  19.150450 0.390031387 0.438137200 0.38709628 0.9948324047 0.750239846
#> 76   5.456840 0.020065218 0.750603328 0.65034390 0.3146626207 0.270977834
#> 77  22.206826 0.376970928 0.667815764 0.81459620 0.8558819662 0.875621909
#> 78  19.830142 0.559912840 0.407973201 0.07096477 0.5403376815 0.851743174
#> 79  17.082682 0.857083586 0.351248815 0.52683032 0.8734173523 0.045849311
#> 80  16.187514 0.384809711 0.738091561 0.76347483 0.5155856195 0.381056859
#> 81  18.561833 0.527917036 0.664285493 0.43538664 0.8694592980 0.183312054
#> 82  14.821010 0.600637524 0.085224700 0.55247234 0.8565225946 0.921799641
#> 83  15.955198 0.261371358 0.856132157 0.20403065 0.3456831840 0.906186849
#> 84   6.082086 0.290050162 0.076983322 0.03102602 0.0001036914 0.192943318
#> 85  20.419140 0.480075174 0.852844803 0.96970706 0.2051797453 0.864375277
#> 86  14.753285 0.920005546 0.106346961 0.17861309 0.9452204907 0.013272602
#> 87  14.598446 0.400720185 0.484802824 0.77829279 0.2813335131 0.885660340
#> 88  15.507339 0.213172711 0.247219110 0.88571080 0.8809342172 0.391542107
#> 89  22.168128 0.671766816 0.686569211 0.83644625 0.9611162767 0.076176313
#> 90   3.745904 0.058614111 0.163623198 0.60536844 0.0040184539 0.646781305
#> 91  15.138609 0.997069135 0.952824800 0.90687946 0.5529002056 0.916653897
#> 92  12.140902 0.149035467 0.321854551 0.03590981 0.2192359727 0.834508620
#> 93  17.240259 0.518556636 0.361534117 0.13141851 0.6467271433 0.501247070
#> 94  20.573750 0.846120055 0.887723417 0.09403037 0.7277020467 0.602144770
#> 95  21.329531 0.718269724 0.828014418 0.69658366 0.8725874424 0.468959034
#> 96   9.180871 0.241314020 0.100656458 0.40572872 0.3819815461 0.886374238
#> 97  25.996613 0.547043368 0.906051578 0.06563664 0.8927457042 0.670548492
#> 98  22.195762 0.834801815 0.772730364 0.12649262 0.8438702857 0.393777719
#> 99  14.290742 0.027956025 0.383370670 0.93733022 0.7297612086 0.551290672
#> 100 12.967050 0.469384300 0.999652457 0.21638023 0.0384149766 0.242492676
#>              x6         x7         x8          x9         x10
#> 1   0.203208829 0.36087369 0.28131306 0.958727899 0.962416305
#> 2   0.068341703 0.65251004 0.70694650 0.359960123 0.111789910
#> 3   0.307273737 0.14355923 0.44218692 0.198986242 0.650485266
#> 4   0.993118094 0.68079862 0.67769852 0.726285317 0.983511552
#> 5   0.116309177 0.85928589 0.70071499 0.168132450 0.179127505
#> 6   0.715423077 0.83862999 0.06840642 0.226481330 0.137970441
#> 7   0.226108731 0.88131292 0.95850352 0.845540055 0.056558586
#> 8   0.133348548 0.23769774 0.24448352 0.345853601 0.546367619
#> 9   0.927281762 0.57863085 0.92628333 0.067319828 0.466504801
#> 10  0.894122657 0.89790251 0.50274708 0.027019454 0.896378375
#> 11  0.203664392 0.34442903 0.93551714 0.118044570 0.018531433
#> 12  0.256982919 0.76085823 0.43879963 0.013695368 0.850974229
#> 13  0.614005604 0.73110365 0.35048654 0.735729368 0.501423423
#> 14  0.441499188 0.85243238 0.47650409 0.923960766 0.166161776
#> 15  0.315594685 0.97581804 0.96615953 0.982190599 0.361006101
#> 16  0.101313761 0.11304750 0.29716193 0.405270204 0.652000184
#> 17  0.272646196 0.97026228 0.29064831 0.269694633 0.874176836
#> 18  0.653699397 0.64808278 0.85972225 0.488288651 0.728305660
#> 19  0.927870102 0.05607304 0.19331267 0.051062453 0.412472856
#> 20  0.026624670 0.47313710 0.80910837 0.002476830 0.159229281
#> 21  0.559491197 0.29462501 0.36548651 0.495834675 0.845703555
#> 22  0.854173652 0.61029962 0.25269965 0.243791283 0.884369102
#> 23  0.075883050 0.12110888 0.85608101 0.255749259 0.395998114
#> 24  0.205828856 0.62941651 0.13002437 0.552318777 0.038140789
#> 25  0.495889647 0.71195530 0.79190766 0.251577628 0.951395967
#> 26  0.261856944 0.61208279 0.35590151 0.628450167 0.783727515
#> 27  0.415961451 0.03442923 0.96434384 0.316115506 0.219439741
#> 28  0.439150000 0.66589079 0.11868321 0.924520392 0.330931987
#> 29  0.182377919 0.65296728 0.45816065 0.906491700 0.177759329
#> 30  0.078328266 0.45185844 0.52365044 0.091969796 0.362516306
#> 31  0.017769956 0.51680410 0.83706173 0.737795152 0.391453242
#> 32  0.728179973 0.67661080 0.03350404 0.569740540 0.903877350
#> 33  0.970845025 0.88093309 0.04332844 0.696488576 0.276456849
#> 34  0.535217588 0.73122915 0.79210294 0.700767504 0.139359318
#> 35  0.166490768 0.39861982 0.20269896 0.039940655 0.451336305
#> 36  0.058175651 0.43052065 0.06879545 0.431522404 0.110329849
#> 37  0.459139019 0.14540735 0.86444107 0.814231365 0.880956980
#> 38  0.622798821 0.45112033 0.07708986 0.566586277 0.533842705
#> 39  0.594819201 0.57878149 0.43100342 0.588247173 0.739250795
#> 40  0.743018237 0.07041712 0.97362256 0.880309032 0.719522785
#> 41  0.524280134 0.74232618 0.57095734 0.719317963 0.264481923
#> 42  0.496825250 0.54922071 0.75440665 0.936428893 0.673856367
#> 43  0.014718803 0.63454427 0.24245433 0.429431614 0.209621585
#> 44  0.957919125 0.13662858 0.62715787 0.286187326 0.690405166
#> 45  0.453089235 0.93023350 0.16841460 0.770643183 0.194759515
#> 46  0.981531849 0.65111864 0.44478963 0.616498637 0.782154128
#> 47  0.544705166 0.52870927 0.16467523 0.159274748 0.015032279
#> 48  0.360042401 0.22511539 0.11827678 0.401598006 0.934172524
#> 49  0.503146105 0.98326687 0.36746439 0.001646489 0.084436133
#> 50  0.074633632 0.10904879 0.94655891 0.417966036 0.034052273
#> 51  0.969791890 0.06994552 0.87541076 0.763934360 0.310959494
#> 52  0.648501314 0.77764551 0.26005891 0.794477667 0.784709724
#> 53  0.984565467 0.65832223 0.65729844 0.242587020 0.280400547
#> 54  0.536133257 0.25920274 0.12186463 0.759170954 0.165283795
#> 55  0.761059405 0.91833784 0.31884804 0.029104153 0.433594813
#> 56  0.564385727 0.14871035 0.62306103 0.437340783 0.814045200
#> 57  0.778454304 0.64068048 0.44699743 0.530409487 0.686513980
#> 58  0.639060989 0.20123842 0.70829345 0.429831351 0.777306163
#> 59  0.378960736 0.74362132 0.89002900 0.809296959 0.091200975
#> 60  0.359530671 0.84763743 0.10430650 0.922564619 0.742623895
#> 61  0.844231416 0.85363587 0.63041172 0.675023474 0.396673976
#> 62  0.456749181 0.80341257 0.84155891 0.093968193 0.296028223
#> 63  0.179034291 0.02387676 0.36426243 0.197463317 0.742924000
#> 64  0.966209196 0.79537771 0.62292278 0.475545125 0.466509440
#> 65  0.713411952 0.88655051 0.46074751 0.714371634 0.773027691
#> 66  0.024886031 0.60003734 0.16341228 0.538079284 0.022117051
#> 67  0.626456494 0.43278817 0.51386686 0.875383010 0.815023199
#> 68  0.827921322 0.92027580 0.14326644 0.293585413 0.303833421
#> 69  0.643069032 0.61601443 0.65029870 0.224765428 0.931795865
#> 70  0.708484262 0.43571816 0.15529453 0.690917831 0.540431014
#> 71  0.870727153 0.89730441 0.50830986 0.312329405 0.254818351
#> 72  0.281812582 0.76901352 0.08587044 0.483298179 0.477587364
#> 73  0.173224842 0.12811576 0.50302666 0.121796640 0.704463991
#> 74  0.109670019 0.53863312 0.03857188 0.690007235 0.007505554
#> 75  0.586222192 0.12901623 0.82403288 0.315731404 0.772408778
#> 76  0.850721327 0.51096488 0.28598557 0.816187683 0.265793564
#> 77  0.921526176 0.54379269 0.24675087 0.545915438 0.971040936
#> 78  0.675661902 0.24177609 0.39799359 0.947317184 0.190525120
#> 79  0.350588162 0.80314412 0.63209288 0.038526407 0.019695889
#> 80  0.008658075 0.94791372 0.94857885 0.647643877 0.612376980
#> 81  0.430898793 0.40736135 0.34011924 0.079619011 0.525731409
#> 82  0.724935097 0.51474456 0.62736599 0.432930770 0.634520042
#> 83  0.412256387 0.98154185 0.82491540 0.760893134 0.395447959
#> 84  0.550862395 0.35418673 0.14511538 0.293758840 0.174868220
#> 85  0.156360570 0.45584015 0.96713913 0.107331893 0.327313162
#> 86  0.043001293 0.22002471 0.08603899 0.072994442 0.614753182
#> 87  0.684644113 0.08183302 0.24046123 0.356314577 0.079166411
#> 88  0.514402497 0.36923607 0.63888373 0.571560843 0.519060970
#> 89  0.945687848 0.12666658 0.50339508 0.677988059 0.306281309
#> 90  0.843096942 0.73987251 0.25052326 0.690719102 0.671395304
#> 91  0.064883093 0.95586629 0.97903377 0.373410821 0.393671740
#> 92  0.762110915 0.66867450 0.65657958 0.097978055 0.747848247
#> 93  0.078384252 0.46808935 0.08984890 0.099009708 0.138926522
#> 94  0.133507663 0.53819245 0.63054293 0.774801583 0.377343046
#> 95  0.809696201 0.33095370 0.87676766 0.151495559 0.641383435
#> 96  0.559383239 0.62868211 0.93028154 0.394606580 0.994194674
#> 97  0.733729686 0.45255247 0.41948585 0.934998404 0.368559058
#> 98  0.349695476 0.88801518 0.49555987 0.739589564 0.337348551
#> 99  0.982668667 0.78420182 0.35270469 0.635668123 0.913685443
#> 100 0.408853880 0.13805955 0.24873487 0.338219850 0.909420361