Friedman benchmark data

Simulate data from the Friedman 1 benchmark problem. These data were originally described in Friedman (1991) and Breiman (1996). For details, see sklearn.datasets.make_friedman1.

Usage

gen_friedman(
  n_samples = 100,
  n_features = 10,
  n_bins = NULL,
  sigma = 0.1,
  seed = NULL
)

Arguments

n_samples

Integer specifying the number of samples (i.e., rows) to generate. Default is 100.

n_features

Integer specifying the number of features to generate. Default is 10.

n_bins

Integer specifying the number of (roughly) equal sized bins to split the response into. Default is NULL for no binning. Setting to a positive integer > 1 effectively turns this into a classification problem where n_bins gives the number of classes.

sigma

Numeric specifying the standard deviation of the noise.

seed

Integer specifying the random seed. If NULL (the default) the results will be different each time the function is run.

References

Breiman, Leo (1996) Bagging predictors. Machine Learning 24, pages 123-140.

Friedman, Jerome H. (1991) Multivariate adaptive regression splines. The Annals of Statistics 19 (1), pages 1-67.

Examples

gen_friedman()
#>             y         x1         x2          x3          x4          x5
#> 1   15.725434 0.63160398 0.49630201 0.233035878 0.107614175 0.953667616
#> 2   12.772419 0.89189219 0.93929023 0.173377940 0.427288506 0.270344988
#> 3   13.308826 0.57883905 0.10845870 0.993523779 0.465172519 0.376173404
#> 4   15.207062 0.85913572 0.84420414 0.303670751 0.671109142 0.020801875
#> 5   18.232848 0.96521578 0.33793557 0.785456419 0.329438033 0.964009105
#> 6   15.787203 0.62937533 0.92171998 0.268711038 0.434458278 0.139783276
#> 7   18.075466 0.62639463 0.81168136 0.288167049 0.628242665 0.217537071
#> 8   12.429673 0.52984309 0.30730614 0.619610041 0.709549429 0.029415131
#> 9   18.704032 0.84350986 0.54347675 0.563484905 0.848612132 0.086528689
#> 10  14.922702 0.91836457 0.99337921 0.273191149 0.951385190 0.309922993
#> 11  25.367842 0.63089681 0.51821782 0.964683459 0.805575485 0.908105232
#> 12   8.127235 0.08618422 0.48423066 0.005243636 0.101645099 0.168447449
#> 13  15.240269 0.70904071 0.29030366 0.505526994 0.688892538 0.460114157
#> 14  16.778828 0.64334325 0.17335249 0.322224151 0.794125709 0.941769748
#> 15  19.935126 0.93729497 0.46075520 0.792227878 0.399591187 0.880822036
#> 16  12.729252 0.94915399 0.68613994 0.732199095 0.211828785 0.139984933
#> 17  20.509833 0.47964078 0.63861278 0.410362843 0.835308114 0.783070858
#> 18  26.989579 0.72611435 0.71610299 0.974773980 0.797955461 0.921193493
#> 19  11.112636 0.35151814 0.39190375 0.660274825 0.518990724 0.252405355
#> 20   8.894075 0.17720366 0.60807056 0.508495697 0.159847355 0.792368156
#> 21  14.164169 0.77758194 0.02750646 0.614244684 0.885169278 0.889032828
#> 22   8.472984 0.75409463 0.03464582 0.125788612 0.472700354 0.022278757
#> 23  12.101767 0.04427024 0.40854438 0.235987209 0.767166789 0.497003303
#> 24   9.147097 0.34885999 0.23240100 0.218026535 0.024680495 0.975253299
#> 25  10.518645 0.57489733 0.84814794 0.399594530 0.014791945 0.040595258
#> 26   6.575578 0.02054167 0.95510378 0.884228823 0.030752620 0.544973019
#> 27  19.891166 0.45687407 0.74960839 0.347189416 0.585822995 0.904512665
#> 28  19.262044 0.51539058 0.85806980 0.417842581 0.470281043 0.879998600
#> 29   9.755289 0.94347576 0.76139702 0.761438500 0.004683771 0.151492427
#> 30  17.523333 0.41308270 0.65127155 0.219071113 0.554227204 0.611171665
#> 31   4.793036 0.11106083 0.26379258 0.344426516 0.175310636 0.317302604
#> 32  14.536216 0.60936711 0.74444278 0.667894086 0.090906075 0.655443201
#> 33   9.674930 0.61198943 0.30863188 0.738116564 0.171005793 0.252507399
#> 34  17.019958 0.86789318 0.74313879 0.174908285 0.189715579 0.802788857
#> 35  18.311786 0.54452629 0.40602412 0.004778256 0.249741022 0.864834490
#> 36  15.364393 0.91121089 0.83588246 0.221218974 0.455197080 0.494336536
#> 37  18.025859 0.25211816 0.88318869 0.189776019 0.510923117 0.928268604
#> 38  19.815715 0.34276031 0.76345673 0.501187273 0.828053105 0.827325471
#> 39   7.143433 0.02330894 0.46576849 0.961872283 0.143787565 0.213251358
#> 40  20.499832 0.98843153 0.26467047 0.946342748 0.515606303 0.811642908
#> 41  15.882821 0.58573002 0.27871853 0.508456009 0.657407577 0.914597115
#> 42   8.250084 0.31685328 0.44230520 0.661982830 0.191292643 0.324574390
#> 43  12.871793 0.99742029 0.24037164 0.934965683 0.120002128 0.229348232
#> 44  14.771365 0.34413282 0.51556427 0.880828817 0.326058107 0.656452741
#> 45  16.979788 0.13621499 0.70056328 0.110008522 0.993680665 0.261841838
#> 46  13.966780 0.01359954 0.01129440 0.002116221 0.858067009 0.092304360
#> 47  21.922583 0.58384127 0.79389365 0.204129484 0.707846627 0.611997532
#> 48   6.762931 0.28937849 0.06498819 0.534323215 0.277037689 0.692802861
#> 49   7.955421 0.40305204 0.17198492 0.569004543 0.324270212 0.542156471
#> 50  11.982507 0.37004308 0.27249323 0.034196666 0.313874153 0.250607523
#> 51  15.703508 0.21593660 0.74100514 0.530365465 0.925238261 0.317171498
#> 52   8.643277 0.10843448 0.57806372 0.945872809 0.233351864 0.089244800
#> 53  22.741781 0.80844580 0.67998119 0.552348390 0.878046106 0.809752819
#> 54  12.744265 0.46410969 0.35750455 0.900540512 0.147101803 0.636685020
#> 55  12.612635 0.58370097 0.81816828 0.692529915 0.168745031 0.013751439
#> 56  12.649229 0.69241540 0.11701306 0.587002683 0.879491001 0.201226292
#> 57  11.309500 0.60940052 0.08363153 0.069124558 0.356408489 0.483149431
#> 58  22.184544 0.34233892 0.97606995 0.975347073 0.412293312 0.972054103
#> 59  15.420899 0.29761060 0.46938285 0.787604583 0.582685557 0.703702564
#> 60  13.421280 0.33176612 0.67698134 0.720279714 0.581412983 0.032312887
#> 61  15.954662 0.38334207 0.70792888 0.061760298 0.073049742 0.784171613
#> 62   7.700680 0.15344467 0.28980686 0.914869446 0.121183962 0.321039679
#> 63  17.015095 0.57595369 0.36674489 0.790972373 0.867127166 0.084275318
#> 64  16.455721 0.80814551 0.68936527 0.177840085 0.138603567 0.662576700
#> 65  12.630661 0.41257286 0.16704847 0.404368866 0.758638662 0.539093214
#> 66   8.420760 0.02489505 0.51727696 0.539935666 0.410231733 0.772725302
#> 67  14.305692 0.50088121 0.97720997 0.854100885 0.153752528 0.058902174
#> 68  13.695848 0.38666578 0.94483118 0.374840105 0.005291036 0.826203384
#> 69  20.892025 0.55063662 0.85954178 0.750350344 0.810799847 0.300853854
#> 70  17.975222 0.52727812 0.75884391 0.568454934 0.749853393 0.228745858
#> 71  27.417922 0.64152077 0.63953573 0.972188313 0.866850405 0.937600516
#> 72   4.514402 0.07077622 0.95451887 0.379021510 0.062715611 0.276570810
#> 73  10.194946 0.24285621 0.94844479 0.292261346 0.250350854 0.004194274
#> 74  13.228682 0.38444426 0.72746281 0.264953879 0.272427458 0.342947678
#> 75  16.416851 0.71715100 0.94280539 0.146564679 0.144007401 0.797231851
#> 76  16.302025 0.70576878 0.55585097 0.707519397 0.468974275 0.297151387
#> 77  18.007485 0.46195624 0.53281644 0.946815268 0.585449137 0.260240483
#> 78  21.709318 0.98941607 0.74429620 0.012410400 0.804465219 0.284701089
#> 79  13.274216 0.57130976 0.91323946 0.667568226 0.127550605 0.292325219
#> 80  10.395490 0.42581309 0.21910318 0.290055258 0.244803825 0.794128292
#> 81  13.766263 0.93957923 0.92015114 0.033661789 0.248600605 0.551111175
#> 82  19.221383 0.50597376 0.94967815 0.128200079 0.253788041 0.766878160
#> 83  11.815962 0.50278663 0.39142518 0.447548565 0.468379917 0.257711956
#> 84  17.694456 0.82110983 0.77264470 0.596021500 0.513174633 0.657485042
#> 85  17.362094 0.28007188 0.90270939 0.435010598 0.971534720 0.078355653
#> 86  13.040536 0.41986751 0.23910238 0.499030498 0.730151887 0.525572403
#> 87  12.512929 0.32283240 0.44255601 0.363683148 0.732814632 0.090247946
#> 88  10.543836 0.04284476 0.38681440 0.800136309 0.332868739 0.985069322
#> 89  16.017272 0.50580777 0.28627654 0.434473842 0.963307575 0.365422890
#> 90  14.977557 0.94571717 0.66803099 0.402803067 0.249481117 0.629342119
#> 91  11.194205 0.06852264 0.38719429 0.920903622 0.445962853 0.459348710
#> 92   4.809261 0.19660674 0.42927697 0.485088261 0.040690342 0.366628685
#> 93  14.063255 0.29296319 0.58857251 0.968926954 0.097544140 0.718666698
#> 94  14.629679 0.39869422 0.59018851 0.520241143 0.440652423 0.686410785
#> 95  16.869729 0.40360112 0.03997759 0.985595249 0.738348732 0.873434379
#> 96  13.606195 0.54862502 0.64820442 0.426865327 0.439622334 0.030949883
#> 97  10.353271 0.12030098 0.83170634 0.284920695 0.269697757 0.704201293
#> 98  10.542032 0.25389481 0.69048611 0.580536834 0.391126807 0.245991576
#> 99  11.131038 0.11972791 0.19870937 0.589180456 0.747792089 0.542022190
#> 100 13.856516 0.97979895 0.86745790 0.572966577 0.445810789 0.942586332
#>             x6          x7           x8          x9        x10
#> 1   0.44046569 0.829900560 0.4331574705 0.499190228 0.75709754
#> 2   0.16595720 0.636486454 0.2646829556 0.489156027 0.80688079
#> 3   0.19448263 0.280233754 0.9288746130 0.218212344 0.95019736
#> 4   0.86038933 0.645945038 0.3436320613 0.182254908 0.92300775
#> 5   0.95867331 0.735448056 0.4556266838 0.755559902 0.17597374
#> 6   0.68081706 0.977677902 0.7610954735 0.446853604 0.02632251
#> 7   0.90549030 0.943852187 0.9527509962 0.893792461 0.61751208
#> 8   0.38679867 0.114208427 0.0363120416 0.810689894 0.10660054
#> 9   0.76275896 0.822268993 0.8472968035 0.886358242 0.35360192
#> 10  0.07874307 0.511863515 0.0786744056 0.941175336 0.67389531
#> 11  0.39168723 0.567880244 0.9573781660 0.372251475 0.64209222
#> 12  0.95191127 0.305877875 0.0111445601 0.499978684 0.45671458
#> 13  0.95671972 0.586331384 0.0208432740 0.364993278 0.16867026
#> 14  0.22564061 0.870191680 0.5183818098 0.648251030 0.23704796
#> 15  0.64051021 0.511491111 0.4726328205 0.489647106 0.36064408
#> 16  0.12654922 0.427278012 0.1340290008 0.122436552 0.67361035
#> 17  0.82474242 0.826665880 0.5716529323 0.567136731 0.07138252
#> 18  0.61579538 0.325387695 0.1681996060 0.949995430 0.87371476
#> 19  0.83353977 0.259448674 0.7567233425 0.032235234 0.93700910
#> 20  0.56425224 0.226821193 0.6982210351 0.887774664 0.29433664
#> 21  0.72173869 0.710050215 0.2891104047 0.959481975 0.29256950
#> 22  0.16106726 0.497926504 0.1201768841 0.626327654 0.30677466
#> 23  0.40885327 0.437662699 0.3685989650 0.057971250 0.34614436
#> 24  0.95986174 0.062248959 0.2654765546 0.075953217 0.25775159
#> 25  0.32663404 0.842673271 0.1858627517 0.119039678 0.90555798
#> 26  0.90306560 0.309475396 0.5396044590 0.141721404 0.28394978
#> 27  0.03567864 0.849474667 0.0731993115 0.746772489 0.86485593
#> 28  0.15680724 0.072885830 0.6956684124 0.024105563 0.06652057
#> 29  0.61031235 0.150364077 0.4084564350 0.598081308 0.80594623
#> 30  0.20253837 0.696607191 0.6659760147 0.292863381 0.99876402
#> 31  0.72700891 0.357531175 0.7662719442 0.828249462 0.16948803
#> 32  0.69773550 0.302191552 0.4323906531 0.590413454 0.45931886
#> 33  0.79256799 0.789645298 0.2839513193 0.422372657 0.90694824
#> 34  0.59224132 0.275168921 0.1463990551 0.860617611 0.34563157
#> 35  0.40671584 0.903231279 0.8776325465 0.764780262 0.93765681
#> 36  0.57755339 0.005268169 0.8243542656 0.748250493 0.33683732
#> 37  0.06529897 0.429633375 0.3765830169 0.489332484 0.20370180
#> 38  0.06501535 0.006063428 0.6641301899 0.595244350 0.06455096
#> 39  0.58096140 0.095566290 0.4822617481 0.973731218 0.61900876
#> 40  0.26845222 0.349912994 0.3797567305 0.432467895 0.36894064
#> 41  0.80410019 0.617382447 0.5890156159 0.029738497 0.90676514
#> 42  0.45563918 0.817285138 0.9631846917 0.841101077 0.41760418
#> 43  0.35015871 0.479157111 0.1288977019 0.104042407 0.78076440
#> 44  0.18202664 0.114971835 0.6184877278 0.776672120 0.39820692
#> 45  0.37639277 0.833423722 0.0053823681 0.104756765 0.21972988
#> 46  0.33176750 0.811699925 0.8837181770 0.070014262 0.02205558
#> 47  0.31293655 0.442816646 0.5454743726 0.469620748 0.30351997
#> 48  0.01810937 0.450417519 0.4698118095 0.793429839 0.93705440
#> 49  0.26582917 0.807190841 0.7474068094 0.181230869 0.27168876
#> 50  0.49159740 0.885277939 0.0002021899 0.826625122 0.63915775
#> 51  0.31727611 0.431776018 0.8801678997 0.089908287 0.41678472
#> 52  0.74083453 0.919872096 0.8875738056 0.710366945 0.16634852
#> 53  0.32695145 0.077221258 0.6995026581 0.235137992 0.35790199
#> 54  0.10848153 0.549834253 0.2067458499 0.383669178 0.95155219
#> 55  0.50993326 0.486827806 0.0371030401 0.646063406 0.04138092
#> 56  0.91052077 0.552276372 0.8836578671 0.419110315 0.82620712
#> 57  0.43657954 0.936959477 0.4843844704 0.780743450 0.59055417
#> 58  0.37317405 0.057483512 0.0938911480 0.633663245 0.83691393
#> 59  0.03445030 0.026840670 0.8507457436 0.070626450 0.81420749
#> 60  0.56570283 0.677506472 0.0068162058 0.942107456 0.73839723
#> 61  0.77821453 0.432986126 0.1255004280 0.673369970 0.64480174
#> 62  0.76261263 0.486891396 0.2554685436 0.828349697 0.19172831
#> 63  0.90307876 0.174741291 0.4246292857 0.147744895 0.03691914
#> 64  0.29798001 0.928025848 0.2561675345 0.700492033 0.13205585
#> 65  0.64014655 0.367728842 0.0298147111 0.342659586 0.59534191
#> 66  0.22536072 0.219025363 0.3581252226 0.189862402 0.42750147
#> 67  0.92278594 0.305184218 0.3583658517 0.809650604 0.93561683
#> 68  0.91802591 0.674947234 0.1372350145 0.898869824 0.20587620
#> 69  0.39080656 0.001142248 0.9200873319 0.838223999 0.16343808
#> 70  0.30690681 0.344302908 0.2252022948 0.553611885 0.99262855
#> 71  0.39580552 0.854151863 0.5260471089 0.398971939 0.14425357
#> 72  0.58137987 0.461596576 0.1232098397 0.054832243 0.55960575
#> 73  0.89701796 0.071054361 0.2143805686 0.030413944 0.06879350
#> 74  0.33380959 0.937908919 0.2421422605 0.914664931 0.60459124
#> 75  0.32780081 0.033926095 0.5642038833 0.832487234 0.69032165
#> 76  0.70233127 0.403011337 0.0415333398 0.971239886 0.99540067
#> 77  0.76498061 0.796316136 0.2077869801 0.671113352 0.69420094
#> 78  0.27757461 0.657490812 0.9489930966 0.571997622 0.27050268
#> 79  0.57359949 0.199604138 0.8542308684 0.512222700 0.79461835
#> 80  0.67824473 0.383551775 0.4889845429 0.666834549 0.01856103
#> 81  0.60842539 0.802736151 0.4641496278 0.199858950 0.21591217
#> 82  0.04041924 0.095073852 0.4194450991 0.509349338 0.32957345
#> 83  0.04374387 0.663881024 0.9837651942 0.333037068 0.93272083
#> 84  0.75994890 0.520583106 0.9760412320 0.236500503 0.07666679
#> 85  0.94630841 0.508282186 0.6985811598 0.710560767 0.26430616
#> 86  0.60750010 0.629998200 0.7318156252 0.170937649 0.72698134
#> 87  0.73191363 0.520938261 0.6155300578 0.469288769 0.13147279
#> 88  0.00179346 0.599793122 0.0969845450 0.316960894 0.01450754
#> 89  0.63005310 0.531073788 0.7548914966 0.002121566 0.68565396
#> 90  0.68554637 0.200089724 0.1922426214 0.684610742 0.43514450
#> 91  0.32536210 0.556279072 0.1025488337 0.513625021 0.07312629
#> 92  0.16183649 0.115238446 0.4767395763 0.823996196 0.83363115
#> 93  0.62751903 0.230801827 0.8373391584 0.762883406 0.45177455
#> 94  0.92302496 0.074675303 0.6958822424 0.329260268 0.06424530
#> 95  0.99729693 0.814918682 0.6686532809 0.686689245 0.29654750
#> 96  0.23998160 0.576465367 0.6371335275 0.102834970 0.60997828
#> 97  0.69454683 0.103156217 0.3675505877 0.495943052 0.91141369
#> 98  0.53035470 0.034010267 0.5466891935 0.352052613 0.53005840
#> 99  0.68025008 0.445773093 0.9666095562 0.625746775 0.73954758
#> 100 0.13803334 0.995326297 0.8762625379 0.705759202 0.27674873