Model-Based Variable Importance

Compute model-based variable importance scores for the predictors in a model. (This function is meant for internal use only.)

vi_model(object, ...)

# S3 method for default
vi_model(object, ...)

# S3 method for C5.0
vi_model(object, ...)

# S3 method for constparty
vi_model(object, ...)

# S3 method for earth
vi_model(object, ...)

# S3 method for gbm
vi_model(object, ...)

# S3 method for H2OBinomialModel
vi_model(object, ...)

# S3 method for H2OMultinomialModel
vi_model(object, ...)

# S3 method for H2ORegressionModel
vi_model(object, ...)

# S3 method for lm
vi_model(object, ...)

# S3 method for randomForest
vi_model(object, ...)

# S3 method for RandomForest
vi_model(object, auc = FALSE, ...)

# S3 method for ranger
vi_model(object, ...)

# S3 method for rpart
vi_model(object, ...)

# S3 method for train
vi_model(object, ...)

# S3 method for xgb.Booster
vi_model(object, ...)

Arguments

object
A fitted model object (e.g., a "randomForest" object).

...
Additional optional arguments.

auc
Logical indicating whether or not to compute the AUC-based variable scores described in Janitza et al. (2012). Only available for cforest objects. See varimpAUC for details. Default is FALSE.

Value

A tidy data frame (i.e., a "tibble" object) with two columns: Variable and Importance. For "glm"-like object, an additional column, called Sign, is also included which includes the sign (i.e., POS/NEG) of the original coefficient.

Details

Coming soon!